refactor(control): split gateway.rs along the thread boundary
`control/gateway.rs` held both sides of the control plane's socket in one
4,567-line file: the UI-thread host that decides what a client may do, and
the gateway thread's accept loop, connection session and dispatchers that
put it on the wire, with the egui panel, the frame emitter and the parked
screenshot arc in between. A reviewer asking what a client can make the UI
do had to read all of it.

The thread boundary is the seam the code already had, so the cut follows it:

- `gateway/server.rs` -- the gateway thread. `gateway_run`, `accept_loop`,
  `connection_session`, `dispatch_prepared`, `dispatch_parked_wait`, the
  waiter manager and every response encoder. Never draws.
- `gateway/semantic.rs` -- the frame emitter and the keys it diffs against.
- `gateway/screenshot.rs` -- the park/arm/harvest arc for a read that needs
  a painted frame. The frame service it is armed against, `execute_on_ui`
  and `begin_frame`, stays with the host: both run every request, not only
  the ones that want pixels.
- `gateway/panel.rs` -- the access panel and its two constants, the only
  part of the host that draws.
- `gateway/tests/mod.rs` -- both test modules, as nested `_tests` modules,
  following the shape `app/tests/` settled on.

`gateway.rs` keeps the UI thread's half: options, identities, lifecycle,
local actions and the frame service, 4,142 production lines down to 1,696.

Every body moved unchanged -- proved by a line-multiset comparison, which
found nothing lost but the two `mod tests {` wrappers, one of them re-added
as the `#[cfg(test)] mod tests;` declaration. The only widenings are the
`pub(super)` marks a child module needs to be called by its parent, plus
`pub(super) struct SemanticBaseline` because the host holds one as a field.
`runtime_id_bytes` moved with the server and is re-exported, because
`control/mod.rs` reaches it by the path it has always had.

The seam is checkable rather than asserted: `grep -nE 'egui|eframe'` over
`server.rs` is empty, and `grep -nE 'TcpStream|TcpListener'` over
`gateway.rs` finds only the import and the `TrackedSocket` field it hands
across. No limit, scope, handshake, rate-limiter or response byte changed;
`contract.rs` and the bridge tests are untouched.

Ceiling tightened 4,142 -> 1,696 and the size `!budget` lowered by 2,446.
None of the four siblings takes a baseline entry: at 1,400, 591, 278 and
278 lines they are all under the 1,500-line threshold, so the whole fall is
real rather than a transfer.
