+++ crates/app/src/control/gateway.rs
@@ -2,18 +2,17 @@
 
 use std::{
     collections::{BTreeMap, BTreeSet, VecDeque},
-    io::Write as _,
-    net::{Ipv4Addr, Shutdown, TcpListener, TcpStream},
+    net::{Ipv4Addr, Shutdown, TcpStream},
     path::PathBuf,
     sync::{
-        Arc, Mutex,
-        atomic::{AtomicBool, AtomicUsize, Ordering},
+        Arc,
+        atomic::{AtomicBool, Ordering},
     },
     thread,
     time::{Duration, Instant},
 };
 
-use crossbeam_channel::{Receiver, Sender, TryRecvError, TrySendError, bounded};
+use crossbeam_channel::{Receiver, Sender, bounded};
 use quantick_control::{
     codec::{BoundedCodec, CodecError, FrameRole},
     cursor::{EventCursor, resolve_event_read},
@@ -26,10 +25,7 @@ use quantick_control::{
         BearerToken, CURRENT_PROTOCOL_VERSION, HandshakeGrant, HandshakeReply, ProtocolLimits,
         ProtocolVersionRange, accept_handshake,
     },
-    id::{
-        ConnectionId, EventKind, InstanceId, ModuleId, PermissionId, PrincipalId, ProcessNonce,
-        ProfileId, RequestId,
-    },
+    id::{ConnectionId, InstanceId, PermissionId, PrincipalId, ProcessNonce, ProfileId, RequestId},
     limits::{
         CONTROL_CLIENT_BURST, CONTROL_CLIENT_RATE_PER_SECOND, CONTROL_HANDSHAKE_TIMEOUT_MS,
         CONTROL_MAX_BUFFERED_RESPONSE_SLOTS, CONTROL_MAX_CONNECTIONS,
@@ -43,7 +39,7 @@ use quantick_control::{
         ResponseOutcome, WireU64,
     },
 };
-use serde_json::{Value, json};
+use serde_json::Value;
 
 use crate::{app::QuantickApp, metrics};
 
@@ -56,10 +52,9 @@ use super::{
         UiReadExecution,
     },
     events::EventsReadInput,
+    evidence,
     evidence::{EvidenceStore, RawScreenshot, SessionIdentity},
-    feed::connection_state,
-    interaction::{SelectionIdentity, selection_identity, selection_snapshot},
-    journal::{EventJournal, JournalSignal, NewEvent},
+    journal::{EventJournal, JournalSignal},
     notify::NotificationLimiter,
     registry::ProjectionRegistry,
     trace::{
@@ -73,13 +68,21 @@ use quantick_control_local::discovery::publish_descriptor;
 #[cfg(test)]
 use quantick_control_local::discovery::publish_descriptor_in;
 
+mod panel;
+mod screenshot;
+mod semantic;
+mod server;
+
+use semantic::SemanticBaseline;
+// `runtime_id_bytes` is re-exported: `control/mod.rs` reaches it by path.
+pub(crate) use server::runtime_id_bytes;
+use server::{drain_bounded_since, gateway_main, random_bytes};
+
 const GATEWAY_COMMAND_CAPACITY: usize = 64;
 const GATEWAY_STATUS_CAPACITY: usize = 256;
 const GATEWAY_LIFECYCLE_CAPACITY: usize = 8;
 const GATEWAY_CRITICAL_STATUS_SLOTS_PER_CONNECTION: usize = 2;
 const CONTROL_UI_MAX_STATUS_UPDATES_PER_FRAME: usize = 32;
-const CONTROL_PANEL_DEFAULT_WIDTH_PX: f32 = 520.0;
-const CONTROL_PANEL_SECTION_SPACING_PX: f32 = 6.0;
 const ACCEPT_POLL_MS: u64 = 5;
 /// How often the waiter manager re-checks deadlines and the cancellation
 /// flag when no tick arrives; a parked wait never waits longer than this past
@@ -783,34 +786,6 @@ impl ControlAccess {
         &self.journal
     }
 
-    /// Hand the gateway a rasterised frame the way the window would.
-    ///
-    /// A headless `egui::Context` answers no `ViewportCommand::Screenshot`, so
-    /// the correlation between an image and the scene captured beside it can
-    /// only be proved by supplying the pixels here. Everything downstream —
-    /// the stamping, the point-to-pixel scaling, the encoding, the digest — is
-    /// the shipped path, unchanged.
-    #[cfg(test)]
-    pub(crate) fn publish_screenshot_for_test(
-        &mut self,
-        app: &mut QuantickApp,
-        raw: RawScreenshot,
-    ) {
-        self.accept_screenshot(app, raw);
-    }
-
-    /// Whether a capture is still waiting for the window to be rasterised.
-    #[cfg(test)]
-    pub(crate) fn awaiting_screenshot_for_test(&self) -> usize {
-        self.awaiting_screenshot.len()
-    }
-
-    /// How many bundles this instance is holding.
-    #[cfg(test)]
-    pub(crate) fn retained_evidence_for_test(&self) -> usize {
-        self.evidence.retained()
-    }
-
     /// Whom an object this call produces belongs to, when the call is a
     /// trace replaying a recorded action. `None` on every ordinary call: the
     /// acting actor is the author.
@@ -1122,47 +1097,6 @@ impl ControlAccess {
         }
     }
 
-    /// Whether this window is holding a rasterised frame a read could use.
-    pub(crate) fn has_screenshot(&self) -> bool {
-        self.screenshot.is_some()
-    }
-
-    /// The registered snapshot scopes the configured grant already reaches.
-    pub(crate) fn readable_scopes(&self) -> Vec<quantick_control::id::SnapshotScopeId> {
-        self.contract.readable_scopes(&self.configured_scopes)
-    }
-
-    /// Whether the configured grant permits rasterising the window.
-    ///
-    /// Asked *before* anything arms a rasterise, because taking the picture is
-    /// what raises the notice: a caller that will be refused the scope one
-    /// step later must not first tell the trader their window was captured.
-    /// The indicator only means something if it is never wrong.
-    pub(crate) fn grants_screenshot(&self) -> bool {
-        self.configured_scopes
-            .iter()
-            .any(|permission| permission.as_str() == super::evidence::SCREENSHOT_PERMISSION_ID)
-    }
-
-    /// Ask the window to rasterise itself for the next read that wants it,
-    /// and take the frame if one has already arrived.
-    ///
-    /// The same arming and the same harvest a deferred remote capture gets.
-    /// Both halves, because a local read runs whether or not the gateway is
-    /// enabled while the frame service that normally harvests runs only when
-    /// it is — so a hook that armed but could not harvest would wait out its
-    /// whole budget for a frame sitting in the input queue.
-    pub(crate) fn service_screenshot(
-        &mut self,
-        app: &mut QuantickApp,
-        ctx: &eframe::egui::Context,
-    ) {
-        self.harvest_screenshot(app, ctx);
-        if self.screenshot.is_none() {
-            self.arm_screenshot(ctx);
-        }
-    }
-
     /// One request on the application thread: the authority checks the
     /// gateway made when it arrived, re-made at the moment it runs, and then
     /// the read or the action itself.
@@ -1310,560 +1244,6 @@ impl ControlAccess {
         }
     }
 
-    /// Take the rasterised frame the window was asked for, if it has arrived.
-    ///
-    /// Costs a frame nothing until something arms it: with no capture waiting
-    /// the input scan does not run at all. When one does arrive the trader is
-    /// told, because a picture of their window leaving the process is not
-    /// something that should happen quietly (threat model O-18).
-    fn harvest_screenshot(&mut self, app: &mut QuantickApp, ctx: &eframe::egui::Context) {
-        if !self.screenshot_armed {
-            return;
-        }
-        // Inside the input lock: clone the image's handle and nothing else.
-        // The rows are converted by the closure below, on the response worker
-        // — a 4K framebuffer is eight million pixels, and paying for that here
-        // would blow a 250 microsecond frame budget by two orders of
-        // magnitude every time an agent asks for a picture.
-        let taken = ctx.input(|input| {
-            let pixels_per_point = input.pixels_per_point();
-            input.events.iter().find_map(|event| match event {
-                eframe::egui::Event::Screenshot { image, .. } => {
-                    let image = Arc::clone(image);
-                    Some(RawScreenshot {
-                        width_px: u32::try_from(image.size[0]).unwrap_or(0),
-                        height_px: u32::try_from(image.size[1]).unwrap_or(0),
-                        pixels_per_point,
-                        rgba: super::evidence::ScreenshotPixels::new(move || {
-                            image
-                                .pixels
-                                .iter()
-                                // Unmultiplied: the toolkit stores colours
-                                // premultiplied by alpha and a PNG's are not,
-                                // so a translucent surface would darken on the
-                                // way out if the raw bytes were copied across.
-                                .flat_map(eframe::egui::Color32::to_srgba_unmultiplied)
-                                .collect()
-                        }),
-                    })
-                }
-                _ => None,
-            })
-        });
-        if let Some(raw) = taken {
-            self.accept_screenshot(app, raw);
-        }
-    }
-
-    /// Take one rasterised frame, and tell the person at the window.
-    ///
-    /// The single door for pixels entering the control plane, so the notice
-    /// cannot be bypassed by whatever hands them over — the window's own
-    /// screenshot event today, a test's fixture in the same breath.
-    fn accept_screenshot(&mut self, app: &mut QuantickApp, raw: RawScreenshot) {
-        self.screenshot_armed = false;
-        self.screenshot = Some(raw);
-        app.show_agent_toast(SCREENSHOT_NOTICE.to_owned());
-    }
-
-    /// Run the captures that were waiting for an image, or give up on them,
-    /// and report how much of the frame's request budget they spent.
-    ///
-    /// Returns before touching anything when none is waiting, which is every
-    /// frame of every session where no client asked for a picture.
-    fn serve_awaiting_screenshot(
-        &mut self,
-        app: &mut QuantickApp,
-        generation: u64,
-        ctx: &eframe::egui::Context,
-        frame_started: Instant,
-    ) -> usize {
-        if self.awaiting_screenshot.is_empty() {
-            return 0;
-        }
-        let now = Instant::now();
-        let mut served = 0usize;
-        let mut waiting = std::mem::take(&mut self.awaiting_screenshot);
-        while let Some(request) = waiting.pop_front() {
-            // Revocation applies to work already in flight, and a parked
-            // capture is in flight. Answered here rather than left to park,
-            // because parking is what keeps asking the window to rasterise —
-            // and a revoked connection must not be able to make the trader's
-            // window flash a capture notice for a request that will be refused
-            // anyway.
-            if request.grant_generation != generation
-                || self.revoked_connections.contains(&request.connection_id)
-            {
-                let _ = request.response.try_send(Err(known_error(
-                    codes::PERMISSION_DENIED,
-                    "connection authority was revoked while the capture waited for a frame",
-                    false,
-                )));
-                continue;
-            }
-            // These captures spend the same frame budget the drain does, and
-            // they run before it. Left uncounted, four waiters plus the
-            // drain's own four would put eight projection passes in one frame
-            // against a documented ceiling of four, and the drain would find
-            // its budget already gone. A waiter this frame cannot serve stays
-            // queued and is served by the next one.
-            if served >= CONTROL_UI_MAX_REQUESTS_PER_FRAME
-                || elapsed_us_since(frame_started) > CONTROL_UI_BUDGET_US
-            {
-                self.awaiting_screenshot.push_back(request);
-                self.awaiting_screenshot.extend(waiting);
-                ctx.request_repaint();
-                return served;
-            }
-            // Given up on *before* the deadline, not at it. `execute_on_ui`
-            // refuses an expired request with `control.timeout` as its very
-            // first act, so waiting to the last moment would answer a window
-            // that never presented with a bare timeout — no bundle, no
-            // `screenshot: frame_not_delivered`, and none of the text, events
-            // and configuration that were collectable all along. The grace is
-            // what buys the honest answer room to be built.
-            let give_up_at = request
-                .deadline
-                .checked_sub(Duration::from_millis(CONTROL_SCREENSHOT_GRACE_MS))
-                .unwrap_or(request.deadline);
-            if self.screenshot.is_none() && now < give_up_at {
-                self.arm_screenshot(ctx);
-                self.awaiting_screenshot.push_back(request);
-                continue;
-            }
-            let result = self.execute_on_ui(app, generation, &request);
-            let _ = request.response.try_send(result);
-            served = served.saturating_add(1);
-        }
-        // Nothing is waiting any more, so nothing is owed a rasterise. Left
-        // set, the flag would suppress every future arming for the rest of the
-        // session (see `arm_screenshot`).
-        if self.awaiting_screenshot.is_empty() {
-            self.screenshot_armed = false;
-        }
-        served
-    }
-
-    /// Whether this request has to wait a frame for the window to be
-    /// rasterised, arming the rasterise if so.
-    fn defer_for_screenshot(&mut self, request: &UiRequest, ctx: &eframe::egui::Context) -> bool {
-        if !request.prepared.dispatch.needs_screenshot() || self.screenshot.is_some() {
-            return false;
-        }
-        if self.awaiting_screenshot.len() >= CONTROL_MAX_SCREENSHOT_WAITERS {
-            return false;
-        }
-        self.arm_screenshot(ctx);
-        true
-    }
-
-    /// Ask the window to rasterise itself, and keep asking while someone is
-    /// waiting.
-    ///
-    /// The command is sent every time rather than once per arming. A window
-    /// that is minimised, occluded, or between viewport states can swallow the
-    /// request and produce no event; an arming that latched would then park
-    /// every future capture behind a command nobody would ever send again, for
-    /// the rest of the session. Repeating it costs one viewport command per
-    /// frame, and only while a capture is actually waiting for a picture.
-    fn arm_screenshot(&mut self, ctx: &eframe::egui::Context) {
-        self.screenshot_armed = true;
-        ctx.send_viewport_cmd(eframe::egui::ViewportCommand::Screenshot);
-        ctx.request_repaint();
-    }
-
-    /// Forget every retained bundle and everything staged around one.
-    ///
-    /// One function, called by both teardown paths, so the two cannot disagree
-    /// about what a withdrawal clears — which is exactly the kind of omission
-    /// nobody notices until a screenshot flag outlives the door it came
-    /// through.
-    fn forget_evidence(&mut self) {
-        self.evidence.clear();
-        self.screenshot = None;
-        self.screenshot_armed = false;
-        self.awaiting_screenshot.clear();
-    }
-
-    /// Record the semantic changes since the last frame: tab, focus,
-    /// selection, feed connection and market, replay state. The baseline is
-    /// compared in place — a handful of integer and string comparisons — and
-    /// refreshed only where something changed, so a quiet frame allocates
-    /// nothing; with access disabled nothing runs at all, the journal starts
-    /// when the human opens the door and records changes, not the state it
-    /// found.
-    fn emit_semantic_changes(&mut self, app: &QuantickApp) {
-        let tabs = app.control_tabs();
-        let active = &tabs[app
-            .control_active_tab_index()
-            .min(tabs.len().saturating_sub(1))];
-        let active_tab_id = active.id;
-        let focused_pane_id = active.pane(active.focused_side()).id;
-        let selection = selection_identity(app);
-        let Some(mut baseline) = self.semantic_baseline.take() else {
-            self.semantic_baseline = Some(SemanticBaseline {
-                active_tab_id,
-                focused_pane_id,
-                selection,
-                tabs: tabs.iter().map(tab_key).collect(),
-            });
-            return;
-        };
-        let now = metrics::wall_clock_ms();
-        if baseline.active_tab_id != active_tab_id {
-            baseline.active_tab_id = active_tab_id;
-            self.record_observed(
-                "workspace",
-                "workspace.tab.activated",
-                json!({ "tab_id": active_tab_id.to_string() }),
-                now,
-            );
-        }
-        if baseline.focused_pane_id != focused_pane_id {
-            baseline.focused_pane_id = focused_pane_id;
-            self.record_observed(
-                "workspace",
-                "workspace.focus.changed",
-                json!({
-                    "tab_id": active_tab_id.to_string(),
-                    "pane_id": focused_pane_id.to_string(),
-                }),
-                now,
-            );
-        }
-        if baseline.selection != selection {
-            baseline.selection = selection;
-            // The owned snapshot is built only now, for the event: it is the
-            // same projection the selection scope publishes, so "changed"
-            // means the same thing to the journal and to a capture.
-            self.record_observed(
-                "interaction",
-                "interaction.selection.changed",
-                json!({ "selection": selection_snapshot(app) }),
-                now,
-            );
-        }
-        for tab in tabs {
-            match baseline.tabs.iter_mut().find(|old| old.tab_id == tab.id) {
-                None => {
-                    let key = tab_key(tab);
-                    self.record_observed(
-                        "workspace",
-                        "workspace.tab.opened",
-                        json!({ "tab_id": key.tab_id.to_string(), "feed_id": key.feed_id, "symbol": key.symbol }),
-                        now,
-                    );
-                    baseline.tabs.push(key);
-                }
-                Some(old) => {
-                    if old.feed_id != tab.active.0 || old.symbol != tab.active.1 {
-                        old.feed_id.clone_from(&tab.active.0);
-                        old.symbol.clone_from(&tab.active.1);
-                        self.record_observed(
-                            "feed",
-                            "feed.market.changed",
-                            json!({ "tab_id": tab.id.to_string(), "feed_id": old.feed_id, "symbol": old.symbol }),
-                            now,
-                        );
-                    }
-                    let connection = connection_state(tab.feed_connection);
-                    if old.connection != connection {
-                        old.connection = connection;
-                        self.record_observed(
-                            "feed",
-                            "feed.connection.changed",
-                            json!({ "tab_id": tab.id.to_string(), "state": connection }),
-                            now,
-                        );
-                    }
-                    let replay = replay_key(tab);
-                    if old.replay != replay {
-                        old.replay = replay;
-                        self.record_observed(
-                            "replay",
-                            "replay.state.changed",
-                            json!({
-                                "tab_id": tab.id.to_string(),
-                                "active": replay.is_some(),
-                                "playing": replay.map(|(playing, _)| playing),
-                                "finished": replay.map(|(_, finished)| finished),
-                            }),
-                            now,
-                        );
-                    }
-                    // One comparison for both analysis scopes: it walks two
-                    // slices per pane and allocates only once it has something
-                    // to say.
-                    if !analysis_matches(tab, &old.panes) {
-                        self.record_analysis_changes(tab, &old.panes, now);
-                        old.panes = pane_analysis_keys(tab);
-                    }
-                }
-            }
-        }
-        if baseline.tabs.len() != tabs.len()
-            || baseline
-                .tabs
-                .iter()
-                .any(|old| !tabs.iter().any(|tab| tab.id == old.tab_id))
-        {
-            baseline.tabs.retain(|old| {
-                let open = tabs.iter().any(|tab| tab.id == old.tab_id);
-                if !open {
-                    self.record_observed(
-                        "workspace",
-                        "workspace.tab.closed",
-                        json!({ "tab_id": old.tab_id.to_string() }),
-                        now,
-                    );
-                }
-                open
-            });
-        }
-        self.semantic_baseline = Some(baseline);
-    }
-
-    /// Journal what changed about one tab's indicators and drawings.
-    ///
-    /// Reached only from a frame where [`analysis_matches`] already answered
-    /// `false`, so the allocation here is paid once per change and never on a
-    /// quiet frame. Identity drives the diff: an indicator is its slot, a
-    /// drawing is its id, and a list that merely reordered emits nothing.
-    ///
-    /// A slot is an address, not an identity — the pane hands it out again
-    /// after a remove — so a slot whose kind changed is reported as the
-    /// detach and the attach it really is, never as one instance mutating.
-    fn record_analysis_changes(
-        &mut self,
-        tab: &crate::tab::Tab,
-        stored: &[PaneAnalysisKey],
-        now: i64,
-    ) {
-        let tab_id = tab.id.to_string();
-        for (pane, side) in analysis_panes(tab) {
-            let was = stored.iter().find(|key| key.pane_id == pane.id);
-            let previous_indicators = was.map(|key| key.indicators.as_slice()).unwrap_or(&[]);
-            let previous_drawings = was.map(|key| key.drawings.as_slice()).unwrap_or(&[]);
-            let side = crate::control::PaneSideDto::from(side);
-            let pane_id = pane.id.to_string();
-
-            let indicators = pane.indicators.all();
-            let drawings = pane.drawings.items();
-
-            for view in indicators {
-                match previous_indicators
-                    .iter()
-                    .find(|key| key.slot == view.slot.0)
-                {
-                    None => self.record_indicator_attached(
-                        &tab_id,
-                        &pane_id,
-                        side,
-                        view.slot.0,
-                        view.kind.as_ref(),
-                        now,
-                    ),
-                    // The pane reused the address for another constructor
-                    // between two frames. Two events, because two instances.
-                    Some(key) if *key.kind != *view.kind => {
-                        self.record_indicator_detached(
-                            &tab_id,
-                            &pane_id,
-                            side,
-                            key.slot,
-                            key.kind.as_ref(),
-                            now,
-                        );
-                        self.record_indicator_attached(
-                            &tab_id,
-                            &pane_id,
-                            side,
-                            view.slot.0,
-                            view.kind.as_ref(),
-                            now,
-                        );
-                    }
-                    Some(key) => {
-                        let failing = view.error.is_some() || view.stale.is_some();
-                        if key.failing != failing {
-                            self.record_observed(
-                                "analysis",
-                                "analysis.indicator.compile_state.changed",
-                                json!({
-                                    "tab_id": tab_id,
-                                    "pane_id": pane_id,
-                                    "pane_side": side,
-                                    "slot_id": view.slot.0.to_string(),
-                                    "kind": view.kind.as_ref(),
-                                    "failing": failing,
-                                }),
-                                now,
-                            );
-                        }
-                    }
-                }
-            }
-            for key in previous_indicators {
-                if !indicators.iter().any(|view| view.slot.0 == key.slot) {
-                    self.record_indicator_detached(
-                        &tab_id,
-                        &pane_id,
-                        side,
-                        key.slot,
-                        key.kind.as_ref(),
-                        now,
-                    );
-                }
-            }
-
-            for drawing in drawings {
-                match previous_drawings.iter().find(|key| key.id == drawing.id.0) {
-                    None => self.record_observed(
-                        "analysis",
-                        "analysis.drawing.created",
-                        json!({
-                            "tab_id": tab_id,
-                            "pane_id": pane_id,
-                            "pane_side": side,
-                            "drawing_id": drawing.id.0.to_string(),
-                            "tool_id": drawing.tool.id(),
-                            "author": author_payload(drawing),
-                        }),
-                        now,
-                    ),
-                    Some(key) if !key.matches(drawing) => self.record_observed(
-                        "analysis",
-                        "analysis.drawing.edited",
-                        json!({
-                            "tab_id": tab_id,
-                            "pane_id": pane_id,
-                            "pane_side": side,
-                            "drawing_id": drawing.id.0.to_string(),
-                            "tool_id": drawing.tool.id(),
-                            "locked": drawing.locked,
-                            "hidden": drawing.hidden,
-                            // Presence, never the text: the journal is held to
-                            // the same rule as the wire.
-                            "user_label_present": drawing.name.is_some(),
-                            "author": author_payload(drawing),
-                        }),
-                        now,
-                    ),
-                    Some(_) => {}
-                }
-            }
-            for key in previous_drawings {
-                if !drawings.iter().any(|drawing| drawing.id.0 == key.id) {
-                    self.record_drawing_removed(&tab_id, &pane_id, side, key.id, now);
-                }
-            }
-        }
-
-        // A pane the layout closed takes its whole arrangement with it. Walked
-        // separately because the loop above only visits panes that still
-        // exist, and a client left holding indicators and marks that are gone
-        // is exactly what this journal is for.
-        for key in stored {
-            if analysis_panes(tab).any(|(pane, _side)| pane.id == key.pane_id) {
-                continue;
-            }
-            let side = crate::control::PaneSideDto::from(key.side);
-            let pane_id = key.pane_id.to_string();
-            for indicator in &key.indicators {
-                self.record_indicator_detached(
-                    &tab_id,
-                    &pane_id,
-                    side,
-                    indicator.slot,
-                    indicator.kind.as_ref(),
-                    now,
-                );
-            }
-            for drawing in &key.drawings {
-                self.record_drawing_removed(&tab_id, &pane_id, side, drawing.id, now);
-            }
-        }
-    }
-
-    fn record_indicator_attached(
-        &mut self,
-        tab_id: &str,
-        pane_id: &str,
-        side: crate::control::PaneSideDto,
-        slot: u64,
-        kind: &str,
-        now: i64,
-    ) {
-        self.record_observed(
-            "analysis",
-            "analysis.indicator.attached",
-            json!({
-                "tab_id": tab_id,
-                "pane_id": pane_id,
-                "pane_side": side,
-                "slot_id": slot.to_string(),
-                "kind": kind,
-            }),
-            now,
-        );
-    }
-
-    fn record_indicator_detached(
-        &mut self,
-        tab_id: &str,
-        pane_id: &str,
-        side: crate::control::PaneSideDto,
-        slot: u64,
-        kind: &str,
-        now: i64,
-    ) {
-        self.record_observed(
-            "analysis",
-            "analysis.indicator.detached",
-            json!({
-                "tab_id": tab_id,
-                "pane_id": pane_id,
-                "pane_side": side,
-                "slot_id": slot.to_string(),
-                "kind": kind,
-            }),
-            now,
-        );
-    }
-
-    fn record_drawing_removed(
-        &mut self,
-        tab_id: &str,
-        pane_id: &str,
-        side: crate::control::PaneSideDto,
-        drawing_id: u64,
-        now: i64,
-    ) {
-        self.record_observed(
-            "analysis",
-            "analysis.drawing.removed",
-            json!({
-                "tab_id": tab_id,
-                "pane_id": pane_id,
-                "pane_side": side,
-                "drawing_id": drawing_id.to_string(),
-            }),
-            now,
-        );
-    }
-
-    fn record_observed(&mut self, module: &str, kind: &str, payload: Value, now: i64) {
-        self.journal.record(
-            NewEvent {
-                module_id: ModuleId::new(module).expect("static module ID is valid"),
-                kind: EventKind::new(kind).expect("static event kind is valid"),
-                actor: None,
-                payload,
-            },
-            now,
-        );
-    }
-
     pub fn needs_frame_service(&self) -> bool {
         !matches!(self.state, AccessState::Disabled)
     }
@@ -1983,262 +1363,6 @@ impl ControlAccess {
             OBSERVER_PROFILE_ID
         };
         ProfileId::new(id).expect("static profile ID is valid")
-    }
-
-    pub fn menu_label(&self) -> &'static str {
-        match self.state {
-            AccessState::Disabled => "Local agent access…",
-            AccessState::Enabling => "Local agent access: enabling…",
-            AccessState::Enabled(_) => "Local agent access: on",
-            AccessState::Disabling(_) => "Local agent access: disabling…",
-        }
-    }
-
-    pub fn draw_panel(&mut self, ctx: &eframe::egui::Context) {
-        if !self.show_panel {
-            return;
-        }
-        let mut open = self.show_panel;
-        eframe::egui::Window::new("Local agent access")
-            .id(eframe::egui::Id::new("control_access_panel"))
-            .open(&mut open)
-            .default_width(CONTROL_PANEL_DEFAULT_WIDTH_PX)
-            .resizable(true)
-            .show(ctx, |ui| self.draw_panel_body(ui));
-        self.show_panel = open;
-    }
-
-    fn draw_panel_body(&mut self, ui: &mut eframe::egui::Ui) {
-        ui.label(
-            "Allows configured local tools such as Codex or Claude to read this already-open Quantick window. Granted data may be sent to the model provider used by that tool.",
-        );
-        ui.add_space(CONTROL_PANEL_SECTION_SPACING_PX);
-        let status = match self.state {
-            AccessState::Disabled => "Off",
-            AccessState::Enabling => "Enabling…",
-            AccessState::Enabled(_) if self.grants_annotate() => {
-                "On — reading, and answering on the chart"
-            }
-            AccessState::Enabled(_) => "On — reading only",
-            AccessState::Disabling(_) => "Disabling and revoking clients…",
-        };
-        ui.horizontal(|ui| {
-            ui.strong("Status:");
-            ui.label(status);
-        });
-        // The one trader gesture this surface owns, said where the trader
-        // configures it: a newcomer finds it here, not in a manual.
-        ui.small(format!(
-            "{} marks what is under the pointer; clients read marks through events.",
-            ui.ctx().format_shortcut(&MARK_SHORTCUT)
-        ));
-
-        if let Some(error) = &self.initialization_error {
-            ui.colored_label(eframe::egui::Color32::LIGHT_RED, error);
-        }
-        if let Some(notice) = &self.notice {
-            ui.label(notice);
-        }
-        if let AccessState::Enabled(runtime) = &self.state {
-            ui.monospace(format!(
-                "{}:{} · instance {}",
-                INSTANCE_DESCRIPTOR_HOST, runtime.public.port, runtime.public.instance_id
-            ));
-            ui.small(format!(
-                "Published at {} · descriptor {}",
-                runtime.public.published_at_unix_ms,
-                runtime.public.descriptor_path.display()
-            ));
-        }
-
-        ui.separator();
-        ui.strong("Read scopes for the next connection");
-        let can_edit = matches!(self.state, AccessState::Disabled);
-        for descriptor in self.contract.selectable_permissions().filter(|descriptor| {
-            !is_annotate_permission(&descriptor.id)
-                && !is_cockpit_permission(&descriptor.id)
-                // The trade tier is not offered here at all. It is not a
-                // read scope — this heading promises reading only, and the
-                // cockpit section exists because that same mistake was made
-                // once already — and it is not offered anywhere else either,
-                // because nothing can currently grant it: no profile the
-                // handshake reaches holds it, so a checkbox would take the
-                // trader's tick and change nothing. An honest absence beats
-                // a control that lies about what it does. The section that
-                // grants it belongs to the change that decides some
-                // connection may trade.
-                && !is_trade_permission(&descriptor.id)
-        }) {
-            let mut selected = self.configured_scopes.contains(&descriptor.id);
-            // The description is the label — a first-week user reads "Chart
-            // framing, viewport, and bars", not `observe.chart` — and the ID
-            // stays beside it because it is what a client asks for by name.
-            let label = if descriptor.sensitive {
-                format!("{} · {} (sensitive)", descriptor.description, descriptor.id)
-            } else {
-                format!("{} · {}", descriptor.description, descriptor.id)
-            };
-            ui.add_enabled(can_edit, eframe::egui::Checkbox::new(&mut selected, label));
-            if can_edit {
-                if selected {
-                    self.configured_scopes.insert(descriptor.id.clone());
-                } else {
-                    self.configured_scopes.remove(&descriptor.id);
-                }
-            }
-        }
-        // The tier that writes is a separate decision, said in the words a
-        // trader would use: everything above lets an assistant *read* the
-        // window; everything here lets it put something in it.
-        ui.add_space(CONTROL_PANEL_SECTION_SPACING_PX);
-        ui.strong("Let an assistant answer on the chart");
-        ui.small(
-            "Objects an assistant places are labelled with its name wherever you see them, and \"Remove objects placed for you\" in the object manager takes them all back at once. Nothing here can delete your own drawings or touch a position. Rearranging your charts is the separate grant below.",
-        );
-        for descriptor in self
-            .contract
-            .selectable_permissions()
-            .filter(|descriptor| is_annotate_permission(&descriptor.id))
-        {
-            let mut selected = self.configured_scopes.contains(&descriptor.id);
-            let label = if descriptor.sensitive {
-                format!("{} · {} (sensitive)", descriptor.description, descriptor.id)
-            } else {
-                format!("{} · {}", descriptor.description, descriptor.id)
-            };
-            ui.add_enabled(can_edit, eframe::egui::Checkbox::new(&mut selected, label));
-            if can_edit {
-                if selected {
-                    self.configured_scopes.insert(descriptor.id.clone());
-                } else {
-                    self.configured_scopes.remove(&descriptor.id);
-                }
-            }
-        }
-        // The cockpit tier, in its own section for the reason it is its own
-        // tier: it is a *write* grant, and it was rendering under "Read
-        // scopes" — a checkbox that rearranges the trader's window, presented
-        // as though it only looked at it.
-        ui.add_space(CONTROL_PANEL_SECTION_SPACING_PX);
-        ui.strong("Let an assistant rearrange your charts");
-        ui.small(
-            "Changes which charts are on screen, where they sit and how wide they are — the same things the layout picker does. Nothing here places or removes an object, and nothing here touches a position. A chart put away keeps its drawings, its indicators and its bars, and comes back with them.",
-        );
-        for descriptor in self
-            .contract
-            .selectable_permissions()
-            .filter(|descriptor| is_cockpit_permission(&descriptor.id))
-        {
-            let mut selected = self.configured_scopes.contains(&descriptor.id);
-            let label = if descriptor.sensitive {
-                format!("{} · {} (sensitive)", descriptor.description, descriptor.id)
-            } else {
-                format!("{} · {}", descriptor.description, descriptor.id)
-            };
-            ui.add_enabled(can_edit, eframe::egui::Checkbox::new(&mut selected, label));
-            if can_edit {
-                if selected {
-                    self.configured_scopes.insert(descriptor.id.clone());
-                } else {
-                    self.configured_scopes.remove(&descriptor.id);
-                }
-            }
-        }
-        if !can_edit {
-            ui.small("Disable access before changing scopes; re-enabling rotates the token and requires a new handshake.");
-        }
-
-        ui.separator();
-        match self.state {
-            AccessState::Disabled => {
-                let label = if self.grants_cockpit() {
-                    "Enable access (reading, answering and rearranging)"
-                } else if self.grants_annotate() {
-                    "Enable access (reading and answering)"
-                } else {
-                    "Enable observer access"
-                };
-                if ui
-                    .add_enabled(self.identity.is_some(), eframe::egui::Button::new(label))
-                    .clicked()
-                {
-                    self.enable(ui.ctx());
-                }
-            }
-            AccessState::Enabling => {
-                if ui.button("Cancel and keep access off").clicked() {
-                    self.request_disable();
-                }
-            }
-            AccessState::Enabled(_) => {
-                if ui.button("Disable and revoke all clients").clicked() {
-                    self.request_disable();
-                }
-            }
-            AccessState::Disabling(_) => {
-                ui.spinner();
-            }
-        }
-
-        ui.separator();
-        ui.strong(format!("Connected clients ({})", self.connections.len()));
-        if self.connections.is_empty() {
-            ui.label("No authenticated clients.");
-        } else {
-            let mut revoke = None;
-            for client in self.connections.values() {
-                ui.group(|ui| {
-                    ui.horizontal(|ui| {
-                        ui.strong(&client.client_name);
-                        if ui.button("Revoke").clicked() {
-                            revoke = Some(client.connection_id.clone());
-                        }
-                    });
-                    ui.small(format!(
-                        "requested {} · effective {} · connected {} · last request {}",
-                        client.requested_profile,
-                        client.effective_profile,
-                        client.connected_at_unix_ms,
-                        client
-                            .last_request_at_unix_ms
-                            .map_or_else(|| "none".to_owned(), |at| at.to_string())
-                    ));
-                    ui.small(format!(
-                        "scopes: {}",
-                        client
-                            .effective_scopes
-                            .iter()
-                            .map(ToString::to_string)
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ));
-                });
-            }
-            if let Some(connection_id) = revoke {
-                self.revoke(connection_id);
-            }
-        }
-
-        ui.separator();
-        ui.small(format!(
-            "Last UI drain: {} request(s), {} µs, budget {} µs{}",
-            self.last_drain.processed,
-            self.last_drain.elapsed_us,
-            CONTROL_UI_BUDGET_US,
-            match (self.last_drain.budget_exceeded, self.last_drain.processed) {
-                (true, 0) => " (budget spent before any request ran)",
-                (true, _) => " (budget exceeded by one non-preemptible capture)",
-                (false, _) => "",
-            }
-        ));
-        let projection = self.projections.performance();
-        ui.small(format!(
-            "Projection captures: {} · last {} µs · worst {} µs · budget violations {}",
-            projection.captures,
-            projection.last_capture_us,
-            projection.worst_capture_us,
-            projection.budget_violations
-        ));
     }
 
     /// Enable observer access with the reviewed defaults. The panel button, the
@@ -2572,1996 +1696,5 @@ impl Drop for ControlAccess {
     }
 }
 
-/// What the frame emitter remembers between frames: owned copies of the
-/// values that name a change, refreshed only where one happened, so a quiet
-/// frame compares in place and allocates nothing.
-struct SemanticBaseline {
-    active_tab_id: u64,
-    focused_pane_id: u64,
-    selection: SelectionIdentity,
-    tabs: Vec<TabKey>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct TabKey {
-    tab_id: u64,
-    feed_id: String,
-    symbol: String,
-    connection: &'static str,
-    /// `(playing, finished)` while a replay is linked.
-    replay: Option<(bool, bool)>,
-    /// What the trader has placed on each pane. Rebuilt only when it stops
-    /// matching — see [`analysis_matches`].
-    panes: Vec<PaneAnalysisKey>,
-}
-
-/// One pane's indicators and drawings, as the journal compares them.
-///
-/// This is a baseline, not a projection: it holds the least that can tell a
-/// change apart, so the per-frame comparison touches integers and booleans and
-/// allocates nothing.
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct PaneAnalysisKey {
-    pane_id: u64,
-    /// Which half of the split this was. Kept so a pane the layout *closed*
-    /// can still name its side in the events that retire what it held — by
-    /// then there is no pane left to ask.
-    side: crate::pane::PaneSide,
-    indicators: Vec<IndicatorKey>,
-    drawings: Vec<DrawingKey>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct IndicatorKey {
-    slot: u64,
-    kind: std::sync::Arc<str>,
-    /// An evaluation error or a failed hot reload. Either one is what a client
-    /// waiting on a compile is waiting for.
-    failing: bool,
-}
-
-impl IndicatorKey {
-    fn matches(&self, view: &crate::indicators::IndicatorView) -> bool {
-        self.slot == view.slot.0
-            && *self.kind == *view.kind
-            && self.failing == (view.error.is_some() || view.stale.is_some())
-    }
-
-    fn of(view: &crate::indicators::IndicatorView) -> Self {
-        Self {
-            slot: view.slot.0,
-            kind: std::sync::Arc::clone(&view.kind),
-            failing: view.error.is_some() || view.stale.is_some(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct DrawingKey {
-    id: u64,
-    locked: bool,
-    hidden: bool,
-    /// Whether the trader gave it a name — never the name, which is their own
-    /// text and does not belong in the journal any more than on the wire.
-    named: bool,
-    /// `(actor kind, client name)` when something other than the trader's hand
-    /// placed it.
-    author: Option<(String, String)>,
-}
-
-impl DrawingKey {
-    fn matches(&self, drawing: &crate::drawings::Drawing) -> bool {
-        self.id == drawing.id.0
-            && self.locked == drawing.locked
-            && self.hidden == drawing.hidden
-            && self.named == drawing.name.is_some()
-            && match (&self.author, &drawing.author) {
-                (None, None) => true,
-                (Some((kind, name)), Some(author)) => {
-                    kind == &author.actor_kind && name == &author.client_name
-                }
-                (None, Some(_)) | (Some(_), None) => false,
-            }
-    }
-
-    fn of(drawing: &crate::drawings::Drawing) -> Self {
-        Self {
-            id: drawing.id.0,
-            locked: drawing.locked,
-            hidden: drawing.hidden,
-            named: drawing.name.is_some(),
-            author: drawing
-                .author
-                .as_ref()
-                .map(|author| (author.actor_kind.clone(), author.client_name.clone())),
-        }
-    }
-}
-
-/// A drawing's author as the journal carries it. Absent is the trader's own
-/// hand; anything else names who acted, under the same rule that labels an
-/// inferred side.
-fn author_payload(drawing: &crate::drawings::Drawing) -> Value {
-    drawing.author.as_ref().map_or(Value::Null, |author| {
-        json!({
-            "actor_kind": author.actor_kind,
-            "client_name": author.client_name,
-        })
-    })
-}
-
-fn tab_key(tab: &crate::tab::Tab) -> TabKey {
-    TabKey {
-        tab_id: tab.id,
-        feed_id: tab.active.0.clone(),
-        symbol: tab.active.1.clone(),
-        connection: connection_state(tab.feed_connection),
-        replay: replay_key(tab),
-        panes: pane_analysis_keys(tab),
-    }
-}
-
-/// The panes of one tab, in the order both analysis scopes publish them.
-///
-/// [`crate::tab::Tab::panes`] and not a local `Vec`: this is walked by
-/// [`analysis_matches`] on every frame, and the version of this comparison
-/// #223's review removed was removed for allocating exactly one vector per
-/// frame. One walk shared with the scopes also keeps the journal's order and
-/// the wire's order from drifting apart.
-fn analysis_panes(
-    tab: &crate::tab::Tab,
-) -> impl Iterator<Item = (&crate::pane::ChartPane, crate::pane::PaneSide)> {
-    tab.panes()
-}
-
-fn pane_analysis_keys(tab: &crate::tab::Tab) -> Vec<PaneAnalysisKey> {
-    analysis_panes(tab)
-        .map(|(pane, side)| PaneAnalysisKey {
-            pane_id: pane.id,
-            side,
-            indicators: pane.indicators.all().iter().map(IndicatorKey::of).collect(),
-            drawings: pane.drawings.items().iter().map(DrawingKey::of).collect(),
-        })
-        .collect()
-}
-
-/// Whether the stored baseline still describes this tab's panes.
-///
-/// Compared in place and allocating nothing: a quiet frame walks two slices
-/// and returns. Only a frame that answers `false` pays for
-/// [`pane_analysis_keys`] and the events below.
-fn analysis_matches(tab: &crate::tab::Tab, stored: &[PaneAnalysisKey]) -> bool {
-    let mut stored = stored.iter();
-    let matched = analysis_panes(tab).all(|(pane, _side)| {
-        let Some(key) = stored.next() else {
-            return false;
-        };
-        {
-            let indicators = pane.indicators.all();
-            let drawings = pane.drawings.items();
-            pane.id == key.pane_id
-                && indicators.len() == key.indicators.len()
-                && drawings.len() == key.drawings.len()
-                && indicators
-                    .iter()
-                    .zip(&key.indicators)
-                    .all(|(view, stored)| stored.matches(view))
-                && drawings
-                    .iter()
-                    .zip(&key.drawings)
-                    .all(|(drawing, stored)| stored.matches(drawing))
-        }
-    });
-    // Every stored pane must have been consumed too: a closed pane leaves the
-    // iterator short, and a tab that lost one has changed.
-    matched && stored.next().is_none()
-}
-
-fn replay_key(tab: &crate::tab::Tab) -> Option<(bool, bool)> {
-    tab.replay
-        .as_ref()
-        .map(|link| (link.status.is_playing(), link.status.is_finished()))
-}
-
-/// Microseconds spent since `started`, saturating.
-///
-/// One reading for everything that shares the frame budget, so the drain and
-/// the captures that run before it are measuring the same thing.
-fn elapsed_us_since(started: Instant) -> u64 {
-    u64::try_from(started.elapsed().as_micros()).unwrap_or(u64::MAX)
-}
-
-fn drain_bounded_since<T>(
-    receiver: &Receiver<T>,
-    started: Instant,
-    already_processed: usize,
-    mut handle: impl FnMut(T),
-) -> DrainObservation {
-    // Requests this frame already ran elsewhere (the screenshot waiters) count
-    // against the same ceiling: one frame, one budget.
-    let mut processed = already_processed;
-    // Why the drain stopped is recorded where it stops, not re-derived from
-    // a later clock reading: the count ceiling, the budget, or an empty queue
-    // are three different diagnoses.
-    let mut stopped_on_budget = false;
-    loop {
-        if processed >= CONTROL_UI_MAX_REQUESTS_PER_FRAME {
-            break;
-        }
-        if elapsed_us_since(started) >= CONTROL_UI_BUDGET_US {
-            stopped_on_budget = true;
-            break;
-        }
-        match receiver.try_recv() {
-            Ok(request) => {
-                handle(request);
-                processed += 1;
-            }
-            Err(TryRecvError::Empty | TryRecvError::Disconnected) => break,
-        }
-    }
-    let elapsed_us = elapsed_us_since(started);
-    DrainObservation {
-        processed,
-        elapsed_us,
-        budget_exceeded: stopped_on_budget,
-        queue_has_more: !receiver.is_empty(),
-    }
-}
-
-fn gateway_main(
-    start: GatewayStart,
-    contract: Arc<ObserverContract>,
-    lifecycle: Sender<LifecycleEvent>,
-    wake: impl Fn() + Send + Sync + 'static,
-) {
-    let wake: Arc<dyn Fn() + Send + Sync> = Arc::new(wake);
-    if let Err(message) = gateway_run(start.clone(), contract, &lifecycle, wake) {
-        let _ = lifecycle.send(LifecycleEvent::Failed {
-            generation: start.grant_generation,
-            message,
-        });
-    }
-}
-
-fn gateway_run(
-    start: GatewayStart,
-    contract: Arc<ObserverContract>,
-    lifecycle: &Sender<LifecycleEvent>,
-    wake: Arc<dyn Fn() + Send + Sync>,
-) -> Result<(), String> {
-    start.options.validate().map_err(str::to_owned)?;
-    let token = BearerToken::from_bytes(random_bytes::<CONTROL_TOKEN_BYTES>()?);
-    let listener = TcpListener::bind((Ipv4Addr::LOCALHOST, 0))
-        .map_err(|error| format!("Could not bind the private loopback gateway: {error}"))?;
-    listener
-        .set_nonblocking(true)
-        .map_err(|error| format!("Could not configure the loopback gateway: {error}"))?;
-    let address = listener
-        .local_addr()
-        .map_err(|error| format!("Could not read the loopback gateway address: {error}"))?;
-    if address.ip() != Ipv4Addr::LOCALHOST || address.port() == 0 {
-        return Err("Gateway did not bind literal IPv4 loopback; access remains off.".to_owned());
-    }
-    let published_at_unix_ms = metrics::wall_clock_ms();
-    let descriptor = InstanceDescriptor {
-        descriptor_version: INSTANCE_DESCRIPTOR_VERSION,
-        instance_id: start.identity.instance_id.clone(),
-        process_nonce: start.identity.process_nonce.clone(),
-        process_id: std::process::id(),
-        process_started_at_unix_ms: start.identity.process_started_at_unix_ms,
-        application_version: env!("CARGO_PKG_VERSION").to_owned(),
-        application_commit: option_env!("QUANTICK_GIT_COMMIT")
-            .unwrap_or("unknown")
-            .to_owned(),
-        protocol_versions: ProtocolVersionRange::new(
-            CURRENT_PROTOCOL_VERSION,
-            CURRENT_PROTOCOL_VERSION,
-        )
-        .expect("current protocol range is valid"),
-        transport: INSTANCE_DESCRIPTOR_TRANSPORT.to_owned(),
-        host: INSTANCE_DESCRIPTOR_HOST.to_owned(),
-        port: address.port(),
-        bearer_token: token.clone(),
-        published_at_unix_ms,
-    };
-    #[cfg(test)]
-    let published = match &start.options.descriptor_directory {
-        Some(directory) => publish_descriptor_in(directory, &descriptor),
-        None => publish_descriptor(&descriptor),
-    };
-    #[cfg(not(test))]
-    let published = publish_descriptor(&descriptor);
-    let published = published.map_err(|error| {
-        format!("Could not publish private gateway discovery; access remains off: {error}")
-    })?;
-
-    let (request_tx, request_rx) = bounded(start.options.request_queue_capacity);
-    let (status_tx, status_rx) = bounded(GATEWAY_STATUS_CAPACITY);
-    let (command_tx, command_rx) = bounded(GATEWAY_COMMAND_CAPACITY);
-    let runtime = GatewayRuntime {
-        grant_generation: start.grant_generation,
-        requests: request_rx,
-        statuses: status_rx,
-        commands: command_tx.clone(),
-        cancellation: Arc::clone(&start.cancellation),
-        public: GatewayPublicInfo {
-            instance_id: start.identity.instance_id.clone(),
-            port: address.port(),
-            descriptor_path: published.path().to_path_buf(),
-            published_at_unix_ms,
-        },
-    };
-    if lifecycle
-        .send(LifecycleEvent::Started {
-            generation: start.grant_generation,
-            runtime,
-        })
-        .is_err()
-    {
-        let _ = published.remove();
-        return Ok(());
-    }
-    wake();
-    tracing::info!(
-        target: "quantick::control",
-        event_code = "CONTROL_GATEWAY_ENABLED",
-        instance_id = %start.identity.instance_id,
-        port = address.port(),
-        "local observer gateway enabled"
-    );
-
-    let (park_tx, park_rx) = bounded(CONTROL_MAX_PARKED_WAITERS);
-    {
-        let ticks = start.journal_ticks.clone();
-        let signal = Arc::clone(&start.journal_signal);
-        let cancellation = Arc::clone(&start.cancellation);
-        if thread::Builder::new()
-            .name("quantick-control-waiters".to_owned())
-            .spawn(move || waiter_manager(ticks, park_rx, signal, cancellation))
-            .is_err()
-        {
-            // The descriptor is already on disk with this run's bearer token
-            // and port. Leaving it there would advertise a token for a port
-            // the operating system is about to hand to somebody else.
-            let _ = published.remove();
-            return Err(
-                "Could not start the gateway's waiter manager; access remains off.".to_owned(),
-            );
-        }
-    }
-    let authority = Arc::new(ConnectionAuthority {
-        identity: start.identity.clone(),
-        bearer_token: token,
-        profile_ceiling: start.profile_ceiling.clone(),
-        granted_scopes: start.granted_scopes.clone(),
-        grant_generation: start.grant_generation,
-        options: start.options.clone(),
-        contract,
-        requests: request_tx,
-        statuses: status_tx,
-        commands: command_tx,
-        global_in_flight: Arc::new(AtomicUsize::new(0)),
-        cancellation: Arc::clone(&start.cancellation),
-        wake: Arc::clone(&wake),
-        journal_signal: Arc::clone(&start.journal_signal),
-        park: park_tx,
-        parked_waiters: Arc::new(AtomicUsize::new(0)),
-    });
-    accept_loop(listener, command_rx, Arc::clone(&authority));
-
-    if let Err(error) = published.remove() {
-        tracing::warn!(
-            target: "quantick::control",
-            event_code = "CONTROL_DESCRIPTOR_REMOVE_FAILED",
-            error = %error,
-            "could not remove the local gateway descriptor"
-        );
-    }
-    tracing::info!(
-        target: "quantick::control",
-        event_code = "CONTROL_GATEWAY_DISABLED",
-        instance_id = %start.identity.instance_id,
-        "local observer gateway disabled"
-    );
-    let _ = lifecycle.send(LifecycleEvent::Stopped {
-        generation: start.grant_generation,
-    });
-    wake();
-    Ok(())
-}
-
-/// A `wait_for_change` registered with the waiter manager: wake it when the
-/// journal passes `target_sequence`, or at `deadline`.
-struct ParkedWaiter {
-    target_sequence: u64,
-    deadline: Instant,
-    wake: Sender<WakeReason>,
-    /// The connection that parked it: a closed one releases the wait at the
-    /// manager's next pass instead of holding its slots to the deadline.
-    connection: Arc<ConnectionSlots>,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum WakeReason {
-    Woken,
-    TimedOut,
-    Shutdown,
-    Disconnected,
-}
-
-/// What one connection's reader and its response threads share: the
-/// in-flight count and request IDs (contract §5.2), the parked waits it
-/// holds, and whether its socket is still open.
-struct ConnectionSlots {
-    in_flight: AtomicUsize,
-    in_flight_ids: Mutex<BTreeSet<quantick_control::id::RequestId>>,
-    parked: AtomicUsize,
-    closed: AtomicBool,
-}
-
-impl ConnectionSlots {
-    fn new() -> Arc<Self> {
-        Arc::new(Self {
-            in_flight: AtomicUsize::new(0),
-            in_flight_ids: Mutex::new(BTreeSet::new()),
-            parked: AtomicUsize::new(0),
-            closed: AtomicBool::new(false),
-        })
-    }
-
-    /// A poisoned lock reads as "in flight": refusing a request is the safe
-    /// side of a broken invariant.
-    fn is_in_flight(&self, request_id: &quantick_control::id::RequestId) -> bool {
-        self.in_flight_ids
-            .lock()
-            .map(|ids| ids.contains(request_id))
-            .unwrap_or(true)
-    }
-
-    fn track(&self, request_id: &quantick_control::id::RequestId) {
-        if let Ok(mut ids) = self.in_flight_ids.lock() {
-            ids.insert(request_id.clone());
-        }
-    }
-
-    fn forget(&self, request_id: &quantick_control::id::RequestId) {
-        if let Ok(mut ids) = self.in_flight_ids.lock() {
-            ids.remove(request_id);
-        }
-    }
-}
-
-/// The waiter manager: one thread per gateway run that owns the parked
-/// waiters, listens to the journal's tick, and wakes each waiter when its
-/// position is behind the journal or its deadline passed. The application
-/// thread never sees it: it only stores an atomic and tries one send.
-fn waiter_manager(
-    ticks: Receiver<()>,
-    park: Receiver<ParkedWaiter>,
-    signal: Arc<JournalSignal>,
-    cancellation: Arc<AtomicBool>,
-) {
-    let mut waiters: Vec<ParkedWaiter> = Vec::new();
-    loop {
-        let now = Instant::now();
-        let poll = waiters
-            .iter()
-            .map(|waiter| waiter.deadline.saturating_duration_since(now))
-            .min()
-            .unwrap_or(Duration::from_millis(WAITER_POLL_MS))
-            .min(Duration::from_millis(WAITER_POLL_MS));
-        crossbeam_channel::select! {
-            // A disconnected receiver is *always* ready: without this the
-            // journal going away would spin this thread at full speed
-            // instead of ending it.
-            recv(ticks) -> tick => if tick.is_err() { break },
-            recv(park) -> waiter => match waiter {
-                Ok(waiter) => waiters.push(waiter),
-                Err(_) => break,
-            },
-            default(poll) => {}
-        }
-        if cancellation.load(Ordering::Acquire) {
-            for waiter in waiters.drain(..) {
-                let _ = waiter.wake.send(WakeReason::Shutdown);
-            }
-            break;
-        }
-        let next = signal.next_sequence();
-        let now = Instant::now();
-        waiters.retain(|waiter| {
-            if waiter.connection.closed.load(Ordering::Acquire) {
-                let _ = waiter.wake.send(WakeReason::Disconnected);
-                false
-            } else if next > waiter.target_sequence {
-                let _ = waiter.wake.send(WakeReason::Woken);
-                false
-            } else if now >= waiter.deadline {
-                let _ = waiter.wake.send(WakeReason::TimedOut);
-                false
-            } else {
-                true
-            }
-        });
-    }
-    for waiter in waiters.drain(..) {
-        let _ = waiter.wake.send(WakeReason::Shutdown);
-    }
-}
-
-struct ConnectionAuthority {
-    identity: ProcessIdentity,
-    bearer_token: BearerToken,
-    profile_ceiling: ProfileId,
-    granted_scopes: BTreeSet<PermissionId>,
-    grant_generation: u64,
-    options: GatewayOptions,
-    contract: Arc<ObserverContract>,
-    requests: Sender<UiRequest>,
-    statuses: Sender<ConnectionStatus>,
-    commands: Sender<GatewayCommand>,
-    global_in_flight: Arc<AtomicUsize>,
-    cancellation: Arc<AtomicBool>,
-    wake: Arc<dyn Fn() + Send + Sync>,
-    journal_signal: Arc<JournalSignal>,
-    park: Sender<ParkedWaiter>,
-    parked_waiters: Arc<AtomicUsize>,
-}
-
-fn accept_loop(
-    listener: TcpListener,
-    commands: Receiver<GatewayCommand>,
-    authority: Arc<ConnectionAuthority>,
-) {
-    let mut sockets = BTreeMap::<u64, TrackedSocket>::new();
-    let mut next_socket_key = 1u64;
-    let mut shutdown = false;
-    // Capacity rejections answer off this thread; this counts the ones still
-    // in flight so a reconnect loop cannot spawn threads without bound.
-    let rejecting = Arc::new(AtomicUsize::new(0));
-    while !shutdown && !authority.cancellation.load(Ordering::Acquire) {
-        loop {
-            match commands.try_recv() {
-                Ok(GatewayCommand::Shutdown) => {
-                    shutdown = true;
-                    break;
-                }
-                Ok(GatewayCommand::Revoke(connection_id)) => {
-                    for socket in sockets.values() {
-                        if socket.connection_id.as_ref() == Some(&connection_id) {
-                            let _ = socket.stream.shutdown(Shutdown::Both);
-                        }
-                    }
-                }
-                Ok(GatewayCommand::Identified {
-                    socket_key,
-                    connection_id,
-                }) => {
-                    if let Some(socket) = sockets.get_mut(&socket_key) {
-                        socket.connection_id = Some(connection_id);
-                    }
-                }
-                Ok(GatewayCommand::Finished { socket_key }) => {
-                    sockets.remove(&socket_key);
-                }
-                Err(TryRecvError::Empty) => break,
-                Err(TryRecvError::Disconnected) => {
-                    shutdown = true;
-                    break;
-                }
-            }
-        }
-        if shutdown {
-            break;
-        }
-        if authority.cancellation.load(Ordering::Acquire) {
-            break;
-        }
-        loop {
-            match listener.accept() {
-                Ok((stream, peer)) => {
-                    if peer.ip() != Ipv4Addr::LOCALHOST {
-                        let _ = stream.shutdown(Shutdown::Both);
-                        continue;
-                    }
-                    if stream.set_nonblocking(false).is_err() {
-                        let _ = stream.shutdown(Shutdown::Both);
-                        continue;
-                    }
-                    if sockets.len() >= authority.options.max_connections {
-                        // Off the accept thread: the rejection reads the
-                        // client's handshake first (see the function), which
-                        // may wait up to the handshake timeout. Bounded, for
-                        // the same reason the connections are: a peer that
-                        // reconnects in a loop would otherwise spawn a thread
-                        // per attempt, each living to the handshake timeout.
-                        if !try_reserve_in_flight(&rejecting, authority.options.max_connections) {
-                            let _ = stream.shutdown(Shutdown::Both);
-                            continue;
-                        }
-                        let options = authority.options.clone();
-                        let in_flight = Arc::clone(&rejecting);
-                        let spawned = thread::Builder::new()
-                            .name("quantick-control-reject".to_owned())
-                            .spawn(move || {
-                                reject_connection_capacity(stream, &options);
-                                in_flight.fetch_sub(1, Ordering::AcqRel);
-                            });
-                        if spawned.is_err() {
-                            rejecting.fetch_sub(1, Ordering::AcqRel);
-                        }
-                        continue;
-                    }
-                    let Some(next_socket_key_value) = next_socket_key.checked_add(1) else {
-                        let _ = stream.shutdown(Shutdown::Both);
-                        tracing::error!(
-                            target: "quantick::control",
-                            event_code = "CONTROL_SOCKET_ID_EXHAUSTED",
-                            "local gateway exhausted its monotonic socket identity space"
-                        );
-                        shutdown = true;
-                        break;
-                    };
-                    let socket_key = next_socket_key;
-                    next_socket_key = next_socket_key_value;
-                    let tracked = match stream.try_clone() {
-                        Ok(stream) => stream,
-                        Err(_) => {
-                            let _ = stream.shutdown(Shutdown::Both);
-                            continue;
-                        }
-                    };
-                    sockets.insert(
-                        socket_key,
-                        TrackedSocket {
-                            stream: tracked,
-                            connection_id: None,
-                        },
-                    );
-                    let connection_authority = Arc::clone(&authority);
-                    let spawn = thread::Builder::new()
-                        .name(format!("quantick-control-connection-{socket_key}"))
-                        .spawn(move || connection_main(socket_key, stream, connection_authority));
-                    if spawn.is_err()
-                        && let Some(socket) = sockets.remove(&socket_key)
-                    {
-                        let _ = socket.stream.shutdown(Shutdown::Both);
-                    }
-                }
-                Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => break,
-                // A peer that reset or aborted before accept, or a signal, is
-                // that connection's failure, not the listener's: the gateway
-                // stays up for everyone else.
-                Err(error)
-                    if matches!(
-                        error.kind(),
-                        std::io::ErrorKind::ConnectionReset
-                            | std::io::ErrorKind::ConnectionAborted
-                            | std::io::ErrorKind::Interrupted
-                    ) =>
-                {
-                    tracing::debug!(
-                        target: "quantick::control",
-                        event_code = "CONTROL_GATEWAY_ACCEPT_TRANSIENT",
-                        error = %error,
-                        "loopback gateway accept hit a transient error; continuing"
-                    );
-                    continue;
-                }
-                Err(error) => {
-                    tracing::warn!(
-                        target: "quantick::control",
-                        event_code = "CONTROL_GATEWAY_ACCEPT_FAILED",
-                        error = %error,
-                        "loopback gateway accept failed; access is stopping"
-                    );
-                    shutdown = true;
-                    break;
-                }
-            }
-        }
-        if shutdown {
-            break;
-        }
-        thread::sleep(Duration::from_millis(ACCEPT_POLL_MS));
-    }
-    for socket in sockets.values() {
-        let _ = socket.stream.shutdown(Shutdown::Both);
-    }
-}
-
-fn reject_connection_capacity(mut stream: TcpStream, options: &GatewayOptions) {
-    let _ = stream.set_read_timeout(Some(options.handshake_timeout));
-    let _ = stream.set_write_timeout(Some(options.handshake_timeout));
-    let codec = BoundedCodec::handshake();
-    // Read the client's handshake before answering: closing a socket with
-    // unread data resets it, and a reset can discard the rejection before
-    // the client has read it. A frame that never comes or is malformed
-    // changes nothing — the answer is the same.
-    let _ = codec.read_handshake_request(&mut stream);
-    let reply = HandshakeReply::Rejected {
-        error: known_error(
-            codes::BACKPRESSURE,
-            "local gateway connection capacity is full",
-            true,
-        ),
-    };
-    if let Ok(frame) = codec.encode(FrameRole::Response, &reply) {
-        let _ = stream.write_all(&frame);
-    }
-    let _ = stream.shutdown(Shutdown::Both);
-}
-
-fn connection_main(socket_key: u64, mut stream: TcpStream, authority: Arc<ConnectionAuthority>) {
-    let result = connection_session(&mut stream, socket_key, &authority);
-    if let Err(error_code) = result {
-        tracing::debug!(
-            target: "quantick::control",
-            event_code = "CONTROL_CONNECTION_CLOSED",
-            error_code,
-            "local control connection closed"
-        );
-    }
-    let _ = stream.shutdown(Shutdown::Both);
-    let _ = authority
-        .commands
-        .try_send(GatewayCommand::Finished { socket_key });
-}
-
-fn connection_session(
-    stream: &mut TcpStream,
-    socket_key: u64,
-    authority: &Arc<ConnectionAuthority>,
-) -> Result<(), &'static str> {
-    stream
-        .set_read_timeout(Some(authority.options.handshake_timeout))
-        .map_err(|_| codes::AUTH_FAILED)?;
-    stream
-        .set_write_timeout(Some(authority.options.handshake_timeout))
-        .map_err(|_| codes::AUTH_FAILED)?;
-    let handshake_codec = BoundedCodec::handshake();
-    let handshake = match handshake_codec.read_handshake_request(stream) {
-        Ok(request) => request,
-        Err(_) => {
-            send_handshake_rejection(
-                stream,
-                &handshake_codec,
-                known_error(codes::INVALID_REQUEST, "invalid handshake frame", false),
-            );
-            return Err(codes::INVALID_REQUEST);
-        }
-    };
-    let connection_id = ConnectionId::from_bytes(
-        random_bytes::<CONTROL_RUNTIME_ID_BYTES>().map_err(|_| codes::AUTH_FAILED)?,
-    );
-    let principal_id = PrincipalId::from_bytes(
-        random_bytes::<CONTROL_RUNTIME_ID_BYTES>().map_err(|_| codes::AUTH_FAILED)?,
-    );
-    let remote_actor = RemoteActor {
-        principal_id: principal_id.clone(),
-        client_name: handshake.client_name.clone(),
-        connection_id: connection_id.clone(),
-    };
-    let grant = HandshakeGrant {
-        protocol_versions: ProtocolVersionRange::new(
-            CURRENT_PROTOCOL_VERSION,
-            CURRENT_PROTOCOL_VERSION,
-        )
-        .expect("current protocol range is valid"),
-        instance_id: authority.identity.instance_id.clone(),
-        process_nonce: authority.identity.process_nonce.clone(),
-        bearer_token: authority.bearer_token.clone(),
-        connection_id: connection_id.clone(),
-        principal_id,
-        application_version: env!("CARGO_PKG_VERSION").to_owned(),
-        application_commit: option_env!("QUANTICK_GIT_COMMIT")
-            .unwrap_or("unknown")
-            .to_owned(),
-        profile_ceiling: authority.profile_ceiling.clone(),
-        granted_scopes: authority.granted_scopes.clone(),
-        // Advertise the timeout this gateway actually applies, so a client's
-        // own patience is derived from the truth rather than the default.
-        limits: ProtocolLimits {
-            request_timeout_ms: u64::try_from(authority.options.request_timeout.as_millis())
-                .unwrap_or(CONTROL_REQUEST_TIMEOUT_MS),
-            ..ProtocolLimits::default()
-        },
-    };
-    let accepted = match accept_handshake(&handshake, &grant, authority.contract.registry()) {
-        Ok(response) => response,
-        Err(error) => {
-            let code = match error.code.as_str() {
-                codes::AUTH_FAILED => codes::AUTH_FAILED,
-                codes::VERSION_UNSUPPORTED => codes::VERSION_UNSUPPORTED,
-                codes::INVALID_REQUEST => codes::INVALID_REQUEST,
-                _ => codes::PERMISSION_DENIED,
-            };
-            send_handshake_rejection(stream, &handshake_codec, error);
-            tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_AUTHENTICATION_FAILED",
-                error_code = code,
-                "local control authentication failed"
-            );
-            return Err(code);
-        }
-    };
-    // Admission before acceptance: a client told "accepted" and then dropped
-    // for a saturated channel would learn of it only on its first request,
-    // as `control.instance_gone`. Refused here, it hears backpressure.
-    let connected_at_unix_ms = metrics::wall_clock_ms();
-    let client = ConnectedClient {
-        connection_id: connection_id.clone(),
-        client_name: handshake.client_name.clone(),
-        connected_at_unix_ms,
-        requested_profile: handshake.requested_profile.clone(),
-        effective_profile: accepted.effective_profile.clone(),
-        effective_scopes: accepted.effective_scopes.clone(),
-        last_request_at_unix_ms: None,
-    };
-    let saturated = || {
-        known_error(
-            codes::BACKPRESSURE,
-            "local gateway is saturated; retry shortly",
-            true,
-        )
-    };
-    if authority
-        .commands
-        .try_send(GatewayCommand::Identified {
-            socket_key,
-            connection_id: connection_id.clone(),
-        })
-        .is_err()
-    {
-        send_handshake_rejection(stream, &handshake_codec, saturated());
-        return Err(codes::BACKPRESSURE);
-    }
-    if authority.statuses.len() >= activity_status_high_watermark(authority.options.max_connections)
-        || authority
-            .statuses
-            .try_send(ConnectionStatus::Connected(client))
-            .is_err()
-    {
-        send_handshake_rejection(stream, &handshake_codec, saturated());
-        return Err(codes::BACKPRESSURE);
-    }
-    let frame = handshake_codec
-        .encode(
-            FrameRole::Response,
-            &HandshakeReply::Accepted(accepted.clone()),
-        )
-        .map_err(|_| codes::AUTH_FAILED)?;
-    if stream.write_all(&frame).is_err() {
-        // The application already heard "connected": tell it the truth.
-        let _ = authority
-            .statuses
-            .try_send(ConnectionStatus::Disconnected(connection_id.clone()));
-        (authority.wake)();
-        return Err(codes::INSTANCE_GONE);
-    }
-    // One request timeout bounds how long a frame may take to arrive once it
-    // has started; a timeout with nothing received is an idle client and is
-    // not an error (see the read loop below).
-    stream
-        .set_read_timeout(Some(authority.options.request_timeout))
-        .map_err(|_| codes::INSTANCE_GONE)?;
-    stream
-        .set_write_timeout(Some(authority.options.request_timeout))
-        .map_err(|_| codes::INSTANCE_GONE)?;
-
-    let writer_stream = stream.try_clone().map_err(|_| codes::INSTANCE_GONE)?;
-    let writer = Arc::new(Mutex::new(writer_stream));
-    let codec = BoundedCodec::default();
-    // Request IDs in flight on this connection (contract §5.2: a duplicate is
-    // rejected while the first is still executing). Requests that leave for
-    // the application thread and parked waits are in flight from the
-    // reader's point of view; a worker-side read is answered before the next
-    // frame is read.
-    let slots = ConnectionSlots::new();
-    (authority.wake)();
-    tracing::info!(
-        target: "quantick::control",
-        event_code = "CONTROL_CLIENT_CONNECTED",
-        connection_id = %connection_id,
-        client_name = %handshake.client_name,
-        effective_profile = %accepted.effective_profile,
-        "authenticated local control client connected"
-    );
-
-    let mut rate_limiter = ClientRateLimiter::new();
-    loop {
-        let request = match codec.read_request(stream) {
-            Ok(request) => request,
-            // Nothing arrived within one request timeout: an idle client, not
-            // a stalled frame. The connection stays unless the gateway is
-            // going away.
-            Err(CodecError::IdleTimeout) => {
-                if authority.cancellation.load(Ordering::Acquire) {
-                    break;
-                }
-                continue;
-            }
-            // A frame that started and never finished inside the timeout, a
-            // malformed frame, or a closed socket all end the connection: a
-            // half-written frame must not hold a connection thread open.
-            Err(_) => break,
-        };
-        if !rate_limiter.allow(Instant::now()) {
-            send_response(
-                &writer,
-                &codec,
-                failure_response(
-                    &request,
-                    known_error(
-                        codes::BACKPRESSURE,
-                        "client request rate limit is exhausted",
-                        true,
-                    ),
-                ),
-            );
-            tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_CLIENT_RATE_LIMITED",
-                connection_id = %connection_id,
-                "local control client exceeded its request rate"
-            );
-            continue;
-        }
-        if slots.is_in_flight(&request.request_id) {
-            send_response(
-                &writer,
-                &codec,
-                failure_response(
-                    &request,
-                    known_error(
-                        codes::INVALID_REQUEST,
-                        "request_id is already in flight on this connection",
-                        false,
-                    ),
-                ),
-            );
-            continue;
-        }
-        if authority.statuses.len()
-            < activity_status_high_watermark(authority.options.max_connections)
-            && authority
-                .statuses
-                .try_send(ConnectionStatus::Requested {
-                    connection_id: connection_id.clone(),
-                    at_unix_ms: metrics::wall_clock_ms(),
-                })
-                .is_ok()
-        {
-            (authority.wake)();
-        }
-        if request.instance_id != authority.identity.instance_id {
-            send_response(
-                &writer,
-                &codec,
-                failure_response(
-                    &request,
-                    known_error(
-                        codes::INSTANCE_GONE,
-                        "request names a different running instance",
-                        false,
-                    ),
-                ),
-            );
-            continue;
-        }
-        if request.protocol_version != accepted.protocol_version {
-            send_response(
-                &writer,
-                &codec,
-                failure_response(
-                    &request,
-                    known_error(
-                        codes::VERSION_UNSUPPORTED,
-                        "request protocol version differs from the negotiated version",
-                        false,
-                    ),
-                ),
-            );
-            continue;
-        }
-        let prepared = match authority
-            .contract
-            .prepare(request.clone(), &accepted.effective_scopes)
-        {
-            Ok(prepared) => prepared,
-            Err(error) => {
-                send_response(&writer, &codec, failure_response(&request, error));
-                continue;
-            }
-        };
-        dispatch_prepared(
-            prepared,
-            &connection_id,
-            &remote_actor,
-            &accepted,
-            &codec,
-            &writer,
-            &slots,
-            authority,
-        );
-    }
-    // The socket is gone: this connection's parked waits release their slots
-    // at the manager's next pass instead of holding them to the deadline.
-    slots.closed.store(true, Ordering::Release);
-
-    if authority
-        .statuses
-        .try_send(ConnectionStatus::Disconnected(connection_id.clone()))
-        .is_err()
-    {
-        tracing::warn!(
-            target: "quantick::control",
-            event_code = "CONTROL_CONNECTION_STATUS_DROPPED",
-            connection_id = %connection_id,
-            "reserved connection status capacity was unexpectedly exhausted"
-        );
-    }
-    (authority.wake)();
-    tracing::info!(
-        target: "quantick::control",
-        event_code = "CONTROL_CLIENT_DISCONNECTED",
-        connection_id = %connection_id,
-        "local control client disconnected"
-    );
-    Ok(())
-}
-
-fn activity_status_high_watermark(max_connections: usize) -> usize {
-    GATEWAY_STATUS_CAPACITY.saturating_sub(
-        max_connections.saturating_mul(GATEWAY_CRITICAL_STATUS_SLOTS_PER_CONNECTION),
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn dispatch_prepared(
-    prepared: PreparedRequest,
-    connection_id: &ConnectionId,
-    remote_actor: &RemoteActor,
-    handshake: &quantick_control::handshake::HandshakeResponse,
-    codec: &BoundedCodec,
-    writer: &Arc<Mutex<TcpStream>>,
-    slots: &Arc<ConnectionSlots>,
-    authority: &Arc<ConnectionAuthority>,
-) {
-    if let PreparedDispatch::Parked(wait) = &prepared.dispatch {
-        let wait = wait.clone();
-        dispatch_parked_wait(
-            prepared,
-            wait,
-            connection_id,
-            remote_actor,
-            handshake,
-            codec,
-            writer,
-            slots,
-            authority,
-        );
-        return;
-    }
-    // Every terminal path below forgets the request ID: a wait that parked
-    // under this ID is in flight until its read is answered or refused.
-    if !try_reserve_in_flight(
-        &authority.global_in_flight,
-        CONTROL_MAX_BUFFERED_RESPONSE_SLOTS,
-    ) {
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &prepared.envelope,
-                known_error(
-                    codes::BACKPRESSURE,
-                    "global buffered response capacity is full",
-                    true,
-                ),
-            ),
-        );
-        slots.forget(&prepared.envelope.request_id);
-        return;
-    }
-    if let Some(result) = prepared.dispatch.execute_worker(
-        &authority.contract,
-        &authority.identity.instance_id,
-        &handshake.effective_profile,
-        &handshake.effective_scopes,
-        &handshake.effective_limits,
-    ) {
-        let response = serialize_worker_result(&authority.contract, &prepared.envelope, result);
-        send_response(writer, codec, response);
-        authority.global_in_flight.fetch_sub(1, Ordering::AcqRel);
-        slots.forget(&prepared.envelope.request_id);
-        return;
-    }
-    if !try_reserve_in_flight(
-        &slots.in_flight,
-        authority.options.max_in_flight_per_connection,
-    ) {
-        authority.global_in_flight.fetch_sub(1, Ordering::AcqRel);
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &prepared.envelope,
-                known_error(
-                    codes::BACKPRESSURE,
-                    "connection in-flight request capacity is full",
-                    true,
-                ),
-            ),
-        );
-        slots.forget(&prepared.envelope.request_id);
-        return;
-    }
-
-    let envelope = prepared.envelope.clone();
-    slots.track(&envelope.request_id);
-    let (response_tx, response_rx) = bounded(1);
-    let deadline = Instant::now() + authority.options.request_timeout;
-    let response_writer = Arc::clone(writer);
-    let response_codec = codec.clone();
-    let response_slots = Arc::clone(slots);
-    let response_global_in_flight = Arc::clone(&authority.global_in_flight);
-    let contract = Arc::clone(&authority.contract);
-    let wait_envelope = envelope.clone();
-    let spawn = thread::Builder::new()
-        .name(format!("quantick-control-response-{}", envelope.request_id))
-        .spawn(move || {
-            let remaining = deadline.saturating_duration_since(Instant::now());
-            let response = match response_rx.recv_timeout(remaining) {
-                Ok(result) => serialize_ui_result(&contract, &wait_envelope, result),
-                Err(crossbeam_channel::RecvTimeoutError::Timeout) => failure_response(
-                    &wait_envelope,
-                    known_error(
-                        codes::TIMEOUT,
-                        "request did not complete before its deadline",
-                        true,
-                    ),
-                ),
-                Err(crossbeam_channel::RecvTimeoutError::Disconnected) => failure_response(
-                    &wait_envelope,
-                    known_error(
-                        codes::INSTANCE_GONE,
-                        "application request dispatcher is unavailable",
-                        true,
-                    ),
-                ),
-            };
-            send_response(&response_writer, &response_codec, response);
-            response_slots.forget(&wait_envelope.request_id);
-            response_slots.in_flight.fetch_sub(1, Ordering::AcqRel);
-            response_global_in_flight.fetch_sub(1, Ordering::AcqRel);
-        });
-    if spawn.is_err() {
-        slots.forget(&envelope.request_id);
-        slots.in_flight.fetch_sub(1, Ordering::AcqRel);
-        authority.global_in_flight.fetch_sub(1, Ordering::AcqRel);
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &envelope,
-                known_error(
-                    codes::BACKPRESSURE,
-                    "response worker could not be created",
-                    true,
-                ),
-            ),
-        );
-        return;
-    }
-
-    let actor = matches!(prepared.dispatch, PreparedDispatch::Action(_))
-        .then(|| Box::new(remote_actor.context(&prepared.envelope)));
-    let ui_request = UiRequest {
-        prepared,
-        actor,
-        connection_id: connection_id.clone(),
-        grant_generation: authority.grant_generation,
-        deadline,
-        response: response_tx,
-    };
-    match authority.requests.try_send(ui_request) {
-        Ok(()) => {
-            (authority.wake)();
-        }
-        Err(TrySendError::Full(request)) => {
-            let _ = request.response.try_send(Err(known_error(
-                codes::BACKPRESSURE,
-                "application request queue is full",
-                true,
-            )));
-        }
-        Err(TrySendError::Disconnected(request)) => {
-            let _ = request.response.try_send(Err(known_error(
-                codes::INSTANCE_GONE,
-                "application request dispatcher is unavailable",
-                true,
-            )));
-        }
-    }
-}
-
-/// `events.wait`: resolve the position against the journal's published
-/// bounds, answer at once if it is already behind, otherwise park on the
-/// waiter manager — holding one global and one per-connection parked slot
-/// and its request ID, nothing else — and run the bounded read through the
-/// ordinary UI path when woken or timed out.
-#[allow(clippy::too_many_arguments)]
-fn dispatch_parked_wait(
-    prepared: PreparedRequest,
-    wait: ParkedWait,
-    connection_id: &ConnectionId,
-    remote_actor: &RemoteActor,
-    handshake: &quantick_control::handshake::HandshakeResponse,
-    codec: &BoundedCodec,
-    writer: &Arc<Mutex<TcpStream>>,
-    slots: &Arc<ConnectionSlots>,
-    authority: &Arc<ConnectionAuthority>,
-) {
-    let instance_id = authority.identity.instance_id.clone();
-    let position = match resolve_event_read(
-        &instance_id,
-        wait.input.cursor.as_ref(),
-        wait.input.start,
-        authority.journal_signal.bounds(),
-    ) {
-        Ok(position) => position,
-        Err(error) => {
-            send_response(writer, codec, failure_response(&prepared.envelope, error));
-            return;
-        }
-    };
-    let target = position.next_sequence.get();
-    let dropped_before = position.dropped_before;
-    let read_input = EventsReadInput {
-        cursor: Some(EventCursor {
-            instance_id: instance_id.clone(),
-            next_sequence: WireU64::new(target),
-        }),
-        start: None,
-        limit: wait.input.limit,
-    };
-    let envelope = prepared.envelope.clone();
-    let to_read = move |timed_out: bool| PreparedRequest {
-        envelope: prepared.envelope,
-        required_permissions: prepared.required_permissions,
-        dispatch: PreparedDispatch::Ui(Box::new(EventsReadInvocation {
-            input: read_input,
-            timed_out,
-            dropped_before,
-        })),
-    };
-    if authority.journal_signal.next_sequence() > target {
-        // Already behind the journal: no parking, just the read.
-        dispatch_prepared(
-            to_read(false),
-            connection_id,
-            remote_actor,
-            handshake,
-            codec,
-            writer,
-            slots,
-            authority,
-        );
-        return;
-    }
-    if !try_reserve_in_flight(&slots.parked, CONTROL_MAX_PARKED_WAITERS_PER_CONNECTION) {
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &envelope,
-                known_error(
-                    codes::BACKPRESSURE,
-                    "this connection's parked waiter capacity is full",
-                    true,
-                ),
-            ),
-        );
-        return;
-    }
-    if !try_reserve_in_flight(&authority.parked_waiters, CONTROL_MAX_PARKED_WAITERS) {
-        slots.parked.fetch_sub(1, Ordering::AcqRel);
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &envelope,
-                known_error(codes::BACKPRESSURE, "parked waiter capacity is full", true),
-            ),
-        );
-        return;
-    }
-    let (wake_tx, wake_rx) = bounded(1);
-    let deadline = Instant::now() + Duration::from_millis(wait.input.timeout_ms);
-    if authority
-        .park
-        .try_send(ParkedWaiter {
-            target_sequence: target,
-            deadline,
-            wake: wake_tx,
-            connection: Arc::clone(slots),
-        })
-        .is_err()
-    {
-        authority.parked_waiters.fetch_sub(1, Ordering::AcqRel);
-        slots.parked.fetch_sub(1, Ordering::AcqRel);
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &envelope,
-                known_error(codes::BACKPRESSURE, "parked waiter capacity is full", true),
-            ),
-        );
-        return;
-    }
-    // Parked: the ID is in flight until the read is answered (contract §5.2).
-    slots.track(&envelope.request_id);
-    let thread_authority = Arc::clone(authority);
-    let thread_connection_id = connection_id.clone();
-    let thread_remote_actor = remote_actor.clone();
-    let thread_handshake = handshake.clone();
-    let thread_codec = codec.clone();
-    let thread_writer = Arc::clone(writer);
-    let thread_slots = Arc::clone(slots);
-    let thread_envelope = envelope.clone();
-    let spawned = thread::Builder::new()
-        .name("quantick-control-wait".to_owned())
-        .spawn(move || {
-            let reason = wake_rx.recv().unwrap_or(WakeReason::Shutdown);
-            thread_authority
-                .parked_waiters
-                .fetch_sub(1, Ordering::AcqRel);
-            thread_slots.parked.fetch_sub(1, Ordering::AcqRel);
-            match reason {
-                // Nobody is listening: release the ID and write nothing.
-                WakeReason::Disconnected => thread_slots.forget(&thread_envelope.request_id),
-                _ if thread_slots.closed.load(Ordering::Acquire) => {
-                    thread_slots.forget(&thread_envelope.request_id);
-                }
-                WakeReason::Shutdown => {
-                    send_response(
-                        &thread_writer,
-                        &thread_codec,
-                        failure_response(
-                            &thread_envelope,
-                            known_error(
-                                codes::INSTANCE_GONE,
-                                "local access was disabled while the wait was parked",
-                                true,
-                            ),
-                        ),
-                    );
-                    thread_slots.forget(&thread_envelope.request_id);
-                }
-                WakeReason::Woken | WakeReason::TimedOut => dispatch_prepared(
-                    to_read(reason == WakeReason::TimedOut),
-                    &thread_connection_id,
-                    &thread_remote_actor,
-                    &thread_handshake,
-                    &thread_codec,
-                    &thread_writer,
-                    &thread_slots,
-                    &thread_authority,
-                ),
-            }
-        });
-    if spawned.is_err() {
-        // The closure and its wake receiver are gone, so the manager's wake
-        // will fail harmlessly; the slots and the ID are released here and
-        // the client hears the refusal instead of waiting for a reply nobody
-        // would write.
-        authority.parked_waiters.fetch_sub(1, Ordering::AcqRel);
-        slots.parked.fetch_sub(1, Ordering::AcqRel);
-        slots.forget(&envelope.request_id);
-        tracing::warn!(
-            target: "quantick::control",
-            event_code = "CONTROL_WAIT_THREAD_FAILED",
-            "could not create a parked-wait thread"
-        );
-        send_response(
-            writer,
-            codec,
-            failure_response(
-                &envelope,
-                known_error(
-                    codes::BACKPRESSURE,
-                    "parked-wait worker could not be created",
-                    true,
-                ),
-            ),
-        );
-    }
-}
-
-fn try_reserve_in_flight(counter: &AtomicUsize, limit: usize) -> bool {
-    counter
-        .fetch_update(Ordering::AcqRel, Ordering::Acquire, |current| {
-            (current < limit).then_some(current + 1)
-        })
-        .is_ok()
-}
-
-fn serialize_worker_result(
-    contract: &ObserverContract,
-    request: &RequestEnvelope,
-    result: Result<serde_json::Value, ControlError>,
-) -> ResponseEnvelope {
-    match result {
-        Ok(result) => validated_success(contract, request, None, Vec::new(), result),
-        Err(error) => failure_response(request, error),
-    }
-}
-
-fn serialize_ui_result(
-    contract: &ObserverContract,
-    request: &RequestEnvelope,
-    execution: Result<UiReadExecution, ControlError>,
-) -> ResponseEnvelope {
-    match execution {
-        Err(error) => failure_response(request, error),
-        Ok(execution) => match execution.into_serialized() {
-            Ok(serialized) => validated_success(
-                contract,
-                request,
-                serialized.capture_revision,
-                serialized.module_revisions,
-                serialized.result,
-            ),
-            // The reason travels: work that happens after the application
-            // thread can refuse for a reason the client can act on — a bundle
-            // that does not fit its store answers `control.backpressure`, and
-            // whether to retry is a different answer from "serialization
-            // failed".
-            Err(error) => failure_response(request, error),
-        },
-    }
-}
-
-fn validated_success(
-    contract: &ObserverContract,
-    request: &RequestEnvelope,
-    capture_revision: Option<WireU64>,
-    module_revisions: Vec<ModuleRevision>,
-    result: serde_json::Value,
-) -> ResponseEnvelope {
-    let valid =
-        contract.validate_output(&request.capability_id, request.capability_version, &result);
-    if !valid {
-        return failure_response(
-            request,
-            known_error(
-                codes::CAPABILITY_UNAVAILABLE,
-                "observer handler returned data outside its declared schema",
-                false,
-            ),
-        );
-    }
-    ResponseEnvelope {
-        protocol_version: request.protocol_version,
-        request_id: request.request_id.clone(),
-        instance_id: request.instance_id.clone(),
-        capture_revision,
-        module_revisions,
-        outcome: ResponseOutcome::Success { result },
-        warnings: Vec::new(),
-    }
-}
-
-fn failure_response(request: &RequestEnvelope, error: ControlError) -> ResponseEnvelope {
-    ResponseEnvelope {
-        protocol_version: request.protocol_version,
-        request_id: request.request_id.clone(),
-        instance_id: request.instance_id.clone(),
-        capture_revision: None,
-        module_revisions: Vec::new(),
-        outcome: ResponseOutcome::Failure { error },
-        warnings: Vec::new(),
-    }
-}
-
-fn send_handshake_rejection(stream: &mut TcpStream, codec: &BoundedCodec, error: ControlError) {
-    if let Ok(frame) = codec.encode(FrameRole::Response, &HandshakeReply::Rejected { error }) {
-        let _ = stream.write_all(&frame);
-    }
-    let _ = stream.shutdown(Shutdown::Both);
-}
-
-fn send_response(
-    writer: &Arc<Mutex<TcpStream>>,
-    codec: &BoundedCodec,
-    mut response: ResponseEnvelope,
-) {
-    let frame = match codec.encode(FrameRole::Response, &response) {
-        Ok(frame) => frame,
-        Err(error) => {
-            response.capture_revision = None;
-            response.module_revisions.clear();
-            response.outcome = ResponseOutcome::Failure {
-                error: match error {
-                    CodecError::PayloadTooLarge { .. }
-                    | CodecError::StringTooLarge { .. }
-                    | CodecError::JsonTooDeep { .. } => known_error(
-                        codes::PAYLOAD_TOO_LARGE,
-                        "response exceeds the negotiated protocol limit",
-                        false,
-                    ),
-                    _ => known_error(
-                        codes::CAPABILITY_UNAVAILABLE,
-                        "response could not be encoded under the negotiated protocol rules",
-                        false,
-                    ),
-                },
-            };
-            match codec.encode(FrameRole::Response, &response) {
-                Ok(frame) => frame,
-                Err(_) => return,
-            }
-        }
-    };
-    let mut stream = writer
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner());
-    // A write that fails part-way has already put a truncated frame on the
-    // wire: every byte after it would be read as that frame's payload. The
-    // connection cannot be recovered, so it is closed rather than left
-    // writing garbage the client will parse as answers.
-    if stream.write_all(&frame).is_err() {
-        let _ = stream.shutdown(Shutdown::Both);
-    }
-}
-
-fn random_bytes<const N: usize>() -> Result<[u8; N], String> {
-    let mut bytes = [0u8; N];
-    getrandom::fill(&mut bytes).map_err(|error| error.to_string())?;
-    Ok(bytes)
-}
-
-/// Entropy for one runtime identifier, as a control failure rather than a
-/// string.
-///
-/// The evidence reads mint identifiers of their own, and an operating system
-/// that cannot supply entropy has to refuse the request in the vocabulary the
-/// client already handles — never with a guessable identifier.
-pub(crate) fn runtime_id_bytes() -> Result<[u8; CONTROL_RUNTIME_ID_BYTES], ControlError> {
-    random_bytes::<CONTROL_RUNTIME_ID_BYTES>().map_err(|error| {
-        known_error(
-            codes::CAPABILITY_UNAVAILABLE,
-            format!("secure identifier generation failed: {error}"),
-            true,
-        )
-    })
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// A directory that does not exist yet — the only caller asserts exactly
-    /// that — inside this thread's own scratch folder, which is removed when
-    /// the thread ends whether or not anything ever created it.
-    fn unique_test_directory(name: &str) -> PathBuf {
-        static NEXT_DIRECTORY: AtomicUsize = AtomicUsize::new(1);
-        crate::scratch::thread_dir("control-gateway").join(format!(
-            "{name}-{}",
-            NEXT_DIRECTORY.fetch_add(1, Ordering::Relaxed)
-        ))
-    }
-
-    /// The panel may not promise what the contract will refuse.
-    ///
-    /// Every annotate capability requires the `annotate` floor *and* its own
-    /// scope. Granting a scope alone used to raise the ceiling to `annotator`
-    /// and put "answering on the chart" on the status line over a connection
-    /// that was then denied every call it made.
-    #[test]
-    fn granting_an_annotate_scope_grants_the_floor_every_capability_also_needs() {
-        let mut access = ControlAccess::new();
-        access
-            .configure_scopes("all-reads,annotate.chart")
-            .expect("the test grants registered scopes");
-
-        assert!(
-            access.grants_annotate(),
-            "the panel says the window can be answered on"
-        );
-        assert!(
-            access
-                .configured_scopes
-                .iter()
-                .any(|permission| permission.as_str() == ANNOTATE_PERMISSION_ID),
-            "so the floor every annotate capability requires has to be there too"
-        );
-
-        // The contrast: the floor on its own opens nothing, so it must not
-        // make the status line claim the window can be answered on.
-        let mut floor_only = ControlAccess::new();
-        floor_only
-            .configure_scopes("all-reads,annotate")
-            .expect("the test grants registered scopes");
-        assert!(
-            !floor_only.grants_annotate(),
-            "a floor with no scope under it answers on nothing"
-        );
-    }
-
-    #[test]
-    fn empty_discovery_returns_a_next_step_without_creating_or_starting_anything() {
-        let directory = unique_test_directory("empty-discovery");
-        assert!(!directory.exists());
-
-        let discovery = quantick_control_local::client::discover_in(
-            &directory,
-            &quantick_control_local::client::ConnectOptions::observer(
-                "gateway unit test",
-                "0.0.0",
-                BTreeSet::new(),
-            ),
-        )
-        .unwrap();
-        assert!(discovery.clients.is_empty());
-        assert!(discovery.issues.is_empty());
-        assert!(!discovery.next_steps.is_empty());
-        assert!(!directory.exists());
-
-        let error = discovery.select(None).unwrap_err();
-        assert_eq!(error.code.as_str(), codes::INSTANCE_GONE);
-        assert!(!error.context.next_steps.is_empty());
-        assert!(!directory.exists());
-    }
-
-    #[test]
-    fn one_frame_never_drains_more_than_the_reviewed_request_ceiling() {
-        let (sender, receiver) = bounded(CONTROL_UI_MAX_REQUESTS_PER_FRAME + 2);
-        for value in 0..(CONTROL_UI_MAX_REQUESTS_PER_FRAME + 2) {
-            sender.send(value).unwrap();
-        }
-        let mut handled = Vec::new();
-        // A start in the future reads as zero elapsed: the budget cannot
-        // interfere, so this proves the count ceiling alone, whatever the
-        // scheduler does to the test thread.
-        let observation = drain_bounded_since(
-            &receiver,
-            Instant::now() + Duration::from_secs(60),
-            0,
-            |value| handled.push(value),
-        );
-        assert_eq!(handled.len(), CONTROL_UI_MAX_REQUESTS_PER_FRAME);
-        assert_eq!(receiver.len(), 2);
-        assert!(observation.queue_has_more);
-        assert!(
-            !observation.budget_exceeded,
-            "stopping on the count ceiling is not a budget stop"
-        );
-    }
-
-    #[test]
-    fn a_capture_that_exhausts_the_budget_ends_the_frame_drain() {
-        // The count ceiling is the deterministic guard; the elapsed-time
-        // budget is the authoritative one (plan §10.2). Prove the latter on
-        // its own: captures that each cost the whole budget must not run two
-        // to a frame, although the count ceiling alone would allow four. A
-        // scheduler that preempts the thread for longer than the budget
-        // before the first check makes the drain run nothing, which proves
-        // nothing either way — try again, a few times.
-        let budget = Duration::from_micros(CONTROL_UI_BUDGET_US);
-        let mut last = None;
-        for _ in 0..5 {
-            let (sender, receiver) = bounded(4);
-            for value in 0..4 {
-                sender.send(value).unwrap();
-            }
-            let mut handled = 0usize;
-            let observation = drain_bounded_since(&receiver, Instant::now(), 0, |_| {
-                handled += 1;
-                let until = Instant::now() + budget;
-                while Instant::now() < until {
-                    std::hint::spin_loop();
-                }
-            });
-            last = Some((handled, observation, receiver.len()));
-            if handled == 1 {
-                break;
-            }
-        }
-        let (handled, observation, remaining) = last.expect("at least one attempt ran");
-        assert_eq!(
-            handled, 1,
-            "the second capture must wait for the next frame"
-        );
-        assert_eq!(observation.processed, 1);
-        assert!(observation.budget_exceeded);
-        assert!(observation.queue_has_more);
-        assert_eq!(remaining, 3);
-    }
-
-    #[test]
-    fn exit_with_access_disabled_costs_nothing() {
-        // The default state: no gateway thread exists, none will report, and
-        // the application's exit must not wait for one.
-        let mut access = ControlAccess::new();
-        let started = Instant::now();
-        access.shutdown_for_exit();
-        assert!(
-            started.elapsed() < Duration::from_millis(EXIT_SHUTDOWN_TIMEOUT_MS / 4),
-            "a disabled gateway returned at once, not after the exit timeout"
-        );
-    }
-
-    #[test]
-    fn client_rate_limiter_has_a_bounded_burst_and_refills() {
-        let started = Instant::now();
-        let mut limiter = ClientRateLimiter {
-            available_token_nanos: u128::from(CONTROL_CLIENT_BURST)
-                * ClientRateLimiter::ONE_TOKEN_NANOS,
-            last_refill: started,
-        };
-        for _ in 0..CONTROL_CLIENT_BURST {
-            assert!(limiter.allow(started));
-        }
-        assert!(!limiter.allow(started));
-        assert!(limiter.allow(started + Duration::from_secs(1)));
-    }
-
-    #[test]
-    fn global_response_slots_enforce_the_reviewed_buffer_bound() {
-        assert_eq!(
-            CONTROL_MAX_BUFFERED_RESPONSE_SLOTS
-                * quantick_control::limits::CONTROL_MAX_RESPONSE_BYTES,
-            quantick_control::limits::CONTROL_MAX_BUFFERED_RESPONSE_BYTES
-        );
-        let in_flight = AtomicUsize::new(0);
-        for _ in 0..CONTROL_MAX_BUFFERED_RESPONSE_SLOTS {
-            assert!(try_reserve_in_flight(
-                &in_flight,
-                CONTROL_MAX_BUFFERED_RESPONSE_SLOTS
-            ));
-        }
-        assert!(!try_reserve_in_flight(
-            &in_flight,
-            CONTROL_MAX_BUFFERED_RESPONSE_SLOTS
-        ));
-    }
-
-    #[test]
-    fn gateway_options_cannot_exceed_reviewed_hard_limits() {
-        assert!(GatewayOptions::default().validate().is_ok());
-
-        let mut options = GatewayOptions {
-            request_queue_capacity: CONTROL_REQUEST_QUEUE_CAPACITY + 1,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            request_queue_capacity: 0,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            max_connections: CONTROL_MAX_CONNECTIONS + 1,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            max_connections: 0,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            max_in_flight_per_connection: CONTROL_MAX_IN_FLIGHT_PER_CONNECTION + 1,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            max_in_flight_per_connection: 0,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            handshake_timeout: Duration::from_millis(CONTROL_HANDSHAKE_TIMEOUT_MS + 1),
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            handshake_timeout: Duration::ZERO,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            request_timeout: Duration::from_millis(CONTROL_REQUEST_TIMEOUT_MS + 1),
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-
-        options = GatewayOptions {
-            request_timeout: Duration::ZERO,
-            ..GatewayOptions::default()
-        };
-        assert!(options.validate().is_err());
-    }
-
-    #[test]
-    fn a_revoked_connection_stays_revoked_when_its_socket_closes() {
-        // Revoking closes the socket, so the disconnect it causes lands on the
-        // same frame as any request the revoked client had already queued.
-        // `poll_statuses` runs before `drain_bounded_since`, so lifting the
-        // revocation here would hand that request a clean bill of health and
-        // let it act. Connection IDs are random per connection, so keeping the
-        // id costs nothing: no later client can inherit it.
-        let mut access = ControlAccess::new();
-        let (_requests_tx, requests) = bounded(1);
-        let (statuses_tx, statuses) = bounded(1);
-        let (commands, _commands_rx) = bounded(1);
-        let connection_id = ConnectionId::from_bytes([7; 16]);
-        access.state = AccessState::Enabled(GatewayRuntime {
-            grant_generation: 0,
-            requests,
-            statuses,
-            commands,
-            cancellation: Arc::new(AtomicBool::new(false)),
-            public: GatewayPublicInfo {
-                instance_id: InstanceId::from_bytes([1; 16]),
-                port: 0,
-                descriptor_path: PathBuf::new(),
-                published_at_unix_ms: 0,
-            },
-        });
-        access.revoked_connections.insert(connection_id.clone());
-        statuses_tx
-            .try_send(ConnectionStatus::Disconnected(connection_id.clone()))
-            .unwrap();
-        access.poll_statuses(Instant::now());
-        assert!(
-            access.revoked_connections.contains(&connection_id),
-            "the revocation outlives the disconnect it caused"
-        );
-    }
-
-    #[test]
-    fn activity_updates_leave_capacity_for_connection_lifecycle_events() {
-        let high_watermark = activity_status_high_watermark(CONTROL_MAX_CONNECTIONS);
-        assert_eq!(
-            GATEWAY_STATUS_CAPACITY - high_watermark,
-            CONTROL_MAX_CONNECTIONS * GATEWAY_CRITICAL_STATUS_SLOTS_PER_CONNECTION
-        );
-    }
-}
-
-#[cfg(test)]
-mod cockpit_tier_tests {
-    use super::*;
-
-    /// A profile no code path constructs is a tier nothing can reach.
-    ///
-    /// The seven `layout.*` capabilities shipped registered, catalogued and
-    /// refused at the gate, because `configured_profile` knew only the
-    /// observer and the annotator. Nothing failed: the registry was valid, the
-    /// schema published, the catalog complete — and every call was denied. A
-    /// capability set is not delivered until a connection can hold the ceiling
-    /// it needs, so that is what this asserts.
-    #[test]
-    fn granting_the_cockpit_scopes_reaches_the_cockpit_profile() {
-        let mut access = ControlAccess::new();
-        access
-            .configure_scopes("cockpit,cockpit.layout")
-            .expect("the cockpit scopes are registered permissions");
-        assert!(
-            access.grants_cockpit(),
-            "the floor and a scope together open the tier"
-        );
-        assert_eq!(
-            access.configured_profile().as_str(),
-            COCKPIT_PROFILE_ID,
-            "the layout capabilities are refused under any lower ceiling"
-        );
-    }
-
-    /// The floor alone opens nothing, and a scope without the floor opens
-    /// nothing either — every cockpit capability requires both, so claiming
-    /// the tier on half of it would put a grant on the panel that is refused
-    /// on every call.
-    #[test]
-    fn half_the_cockpit_grant_is_not_the_cockpit_tier() {
-        for scopes in ["cockpit", "cockpit.layout"] {
-            let mut access = ControlAccess::new();
-            access
-                .configure_scopes(scopes)
-                .expect("registered permission");
-            assert!(
-                !access.grants_cockpit(),
-                "{scopes} alone must not open the tier"
-            );
-            assert_ne!(access.configured_profile().as_str(), COCKPIT_PROFILE_ID);
-        }
-    }
-
-    /// The annotate tier is untouched by the new one: a trader who granted
-    /// only the chart scopes gets exactly what that consent text describes.
-    #[test]
-    fn the_annotate_tier_does_not_pick_up_the_cockpit_ceiling() {
-        let mut access = ControlAccess::new();
-        access
-            .configure_scopes("annotate,annotate.chart")
-            .expect("registered permissions");
-        assert!(access.grants_annotate());
-        assert!(
-            !access.grants_cockpit(),
-            "annotate must not carry a grant to rearrange the window"
-        );
-        assert_eq!(access.configured_profile().as_str(), ANNOTATOR_PROFILE_ID);
-    }
-
-    /// The tiers have to **nest**, not merely coexist.
-    ///
-    /// `handshake::authorize` intersects the requested and the granted ceiling
-    /// and then demands that one of the two *be* that intersection — "profiles
-    /// are built nested" is the comment it says so under. Two sibling write
-    /// tiers overlap without nesting, and an incomparable pair is refused
-    /// outright: a client asking for `--profile annotator` against a cockpit
-    /// grant could not connect at all.
-    #[test]
-    fn the_cockpit_ceiling_contains_the_annotator_ceiling() {
-        use quantick_control::handshake::ProfileAuthority;
-
-        let access = ControlAccess::new();
-        let registry = access.contract.registry();
-        let ceiling = |id: &str| {
-            registry
-                .permission_ceiling(&ProfileId::new(id).expect("static profile ID is valid"))
-                .expect("the profile is registered")
-        };
-        assert!(
-            ceiling(ANNOTATOR_PROFILE_ID).is_subset(&ceiling(COCKPIT_PROFILE_ID)),
-            "the two write tiers are siblings, so the handshake refuses any \
-             connection that names one against a grant holding the other"
-        );
-    }
-
-    /// A trader who grants both tiers keeps both.
-    ///
-    /// The connection takes one ceiling, and every scope the human ticked has
-    /// to survive it — `authorize` drops a granted scope the ceiling does not
-    /// hold. Under a cockpit ceiling that was a sibling of the annotator, that
-    /// meant ticking "rearrange my charts" silently switched off "answer on
-    /// the chart".
-    #[test]
-    fn granting_both_tiers_keeps_every_scope_the_trader_ticked() {
-        use quantick_control::handshake::ProfileAuthority;
-
-        let mut access = ControlAccess::new();
-        access
-            .configure_scopes("annotate,annotate.chart,cockpit,cockpit.layout")
-            .expect("registered permissions");
-        assert!(access.grants_annotate());
-        assert!(access.grants_cockpit());
-
-        let profile = access.configured_profile();
-        let ceiling = access
-            .contract
-            .registry()
-            .permission_ceiling(&profile)
-            .expect("the configured profile is registered");
-        for scope in &access.configured_scopes {
-            assert!(
-                ceiling.contains(scope),
-                "the {profile} ceiling drops the granted scope {scope}"
-            );
-        }
-    }
-}
+mod tests;
