+++ docs/evidence/gateway-rs-sidecars/report-after.txt
@@ -1,0 +1,58 @@
+crate.lines.app	109311
+crate.lines.backtest	2400
+crate.lines.control	5184
+crate.lines.control-local	1612
+crate.lines.engine	3375
+crate.lines.feed	6707
+crate.lines.feed-binance	2914
+crate.lines.feed-hyperliquid	1773
+crate.lines.feed-mt5	4594
+crate.lines.guards	5559
+crate.lines.indicators	4094
+crate.lines.mcp	2422
+crate.lines.orderbook	1022
+crate.lines.orderflow	7701
+crate.lines.pine	6123
+crate.lines.replay	3787
+crate.lines.sim	2279
+crate.lines.strategy	1879
+crate.lines.trading	1789
+crate.lines.total	174525
+crate.lines.app_percent	62
+file.largest.crates/app/src/drawings/mod.rs	2283
+file.largest.crates/app/src/footprint_render.rs	2236
+file.largest.crates/app/src/orderflow_render.rs	3075
+file.largest.crates/app/src/orderflow_view.rs	2485
+file.largest.crates/app/src/pane.rs	5618
+file.largest.crates/app/src/paper_report.rs	3300
+file.largest.crates/app/src/paper_trading.rs	4589
+file.largest.crates/app/src/tab.rs	4472
+struct.wide.app::ChartPane	77
+struct.wide.app::PaperTrading	31
+struct.wide.app::QuantickApp	56
+struct.wide.app::Tab	62
+struct.wide.app::ToolRail	32
+struct.wide.app::ToolbarModel	31
+struct.wide.orderflow::BookEngine	31
+struct.wide.orderflow::BubbleStyle	30
+struct.wide.orderflow::HeatmapConfig	34
+struct.wide.orderflow::OrderflowHealth	32
+struct.wide.sim::PerformanceReport	36
+ratchet.size.budget	48550
+ratchet.size.recorded	48550
+ratchet.size.measured	174525
+ratchet.context.budget	235924
+ratchet.context.recorded	163195
+ratchet.context.measured	237465
+ratchet.cycle.budget	3
+ratchet.cycle.recorded	3
+ratchet.cycle.measured	3
+site.allow	89
+site.process_id	12
+site.cfg_test	624
+app.lines.without_egui	105603
+graph.edges	25
+headless.findings	0
+scan.unreadable	0
+scan.undecodable	0
+scan.blind	0
